% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{accumulate_data}
\alias{accumulate_data}
\title{Accumulate Data Over Time and/or Space}
\usage{
accumulate_data(d, time.col = "Time", over.time = TRUE,
  longitudinal = TRUE, direction = c("upstream", "downstream"))
}
\arguments{
\item{d}{The data frame containing values to accumulate.}

\item{time.col}{The column containing times.}

\item{over.time}{If \code{TRUE}, accumulate data across time steps. Unless
working with cross-section volume change, this is only valid for data
output at the computation time step.}

\item{longitudinal}{If \code{TRUE}, accumulate data along the reach. Only
valid when all cross sections are included in \code{d}.}

\item{direction}{Accumulate data in the downstream (descending order of cross
section IDs) or upstream (ascending order) direction. Ignored if
\code{longitudinal} is \code{FALSE}.}
}
\value{
A data frame containing the accumulated data from \code{d}. Note
that \code{d} may be reordered in time and by cross-section.
}
\description{
Accumulate data over time and/or longitudinally.
}
