% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{area_to_volume}
\alias{area_to_volume}
\title{Area to Volume}
\usage{
area_to_volume(d, time.col = "Time", station.lengths = NULL)
}
\arguments{
\item{d}{The area data, e.g. output of \code{xs_area}.}

\item{time.col}{The time column name.}

\item{station.lengths}{The cross section length values.}
}
\value{
A wide-format table of cross section volumes.
}
\description{
Compute cross section volumes given cross section areas and lengths.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")

quasi.xs = read_xs(simple.quasi)
quasi.lengths = xs_lengths(read_station_lengths(simple.quasi))
quasi.area = xs_area(quasi.xs)
area_to_volume(quasi.area, station.lengths = quasi.lengths)

}
