% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{change_sediment}
\alias{change_sediment}
\title{Change Over Time (Sediment)}
\usage{
change_sediment(d, time.col = "Time", grain.col = "GrainClass")
}
\arguments{
\item{d}{A wide format data table containing values to compute change
over time.}

\item{time.col}{The time column name.}

\item{grain.col}{the grain class column name.}
}
\value{
A wide-format table of change over time. A value of 0 is assigned to
  the first time step.
}
\description{
Compute change over time from a sediment data table.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.voloutcum = read_sediment(simple.quasi, "Vol Out Cum")

change_sediment(quasi.voloutcum)

}
