% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{change_table}
\alias{change_table}
\title{Change Over Time}
\usage{
change_table(d, time.col = "Time")
}
\arguments{
\item{d}{A wide format data table containing values to compute change
over time.}

\item{time.col}{The time column name.}
}
\value{
A wide-format table of change over time. A value of 0 is assigned to
  the first time step.
}
\description{
Compute change over time from a table.
}
