% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{cumulative_sediment}
\alias{cumulative_sediment}
\title{Accumulate Sediment Data Over Time and/or Space}
\usage{
cumulative_sediment(d, time.col = "Time", grain.col = "GrainClass",
  over.time = TRUE, longitudinal = TRUE, direction = c("upstream",
  "downstream"))
}
\arguments{
\item{d}{A wide-format table containing values to accumulate.}

\item{time.col}{The time column name.}

\item{grain.col}{the grain class column name.}

\item{over.time}{If \code{TRUE}, accumulate data across time steps. This
is generally valid only for data output at the computation time step.}

\item{longitudinal}{If \code{TRUE}, accumulate data along the reach. This
is generally only valid when all cross sections are included in \code{d}.}

\item{direction}{Accumulate data in the downstream (descending order of cross
section IDs) or upstream (ascending order) direction. Ignored if
\code{longitudinal} is \code{FALSE}.}
}
\value{
A data frame containing the accumulated data from \code{d}. Note
  that \code{d} may be reordered in time and by cross-section and grain
  class.
}
\description{
Accumulate data from a sediment table over time and/or longitudinally.
}
