% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{diff_cmi}
\alias{diff_cmi}
\title{Cumulative Mass In Difference Table}
\usage{
diff_cmi(d1, d2, tcol = "Time", gcol = "GrainClass")
}
\arguments{
\item{d1}{The first dataframe.}

\item{d2}{The second dataframe.}

\item{tcol}{the time column name.}

\item{gcol}{the grain class column name.}
}
\value{
A dataframe.
}
\description{
Compute a difference table for cumulative mass in data.
}

