% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{diff_lcmc}
\alias{diff_lcmc}
\title{Longitudinal Cumulative Mass Change Difference Table}
\usage{
diff_lcmc(d1, d2, tcol = "Time", gcol = "GrainClass")
}
\arguments{
\item{d1}{The first dataframe.}

\item{d2}{The second dataframe.}

\item{tcol}{the time column name.}

\item{gcol}{the grain class column name.}
}
\value{
A dataframe.
}
\description{
Compute a difference table for longitudinal cumulative mass change data.
}

