% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{diff_sediment}
\alias{diff_lcmc}
\alias{diff_mic}
\alias{diff_sediment}
\title{Difference Table (Sediment)}
\usage{
diff_sediment(d1, d2, tcol, gcol, diffcol)

diff_lcmc(d1, d2, tcol = "Time", gcol = "GrainClass")

diff_mic(d1, d2, tcol = "Time", gcol = "GrainClass")
}
\arguments{
\item{d1}{The first dataframe.}

\item{d2}{The second dataframe.}

\item{tcol}{the time column name.}

\item{gcol}{the grain class column name.}

\item{diffcol}{The name of the difference column to be created.}
}
\value{
A dataframe.
}
\description{
Compute a difference table from sediment data.
}
\section{Functions}{
\itemize{
\item \code{diff_lcmc}: Compute a difference table for Longitudinal Cumulative Mass Change data.

\item \code{diff_mic}: Compute a difference table for Mass In Cumulative data.
}}

