% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r, R/main.r
\name{diff_dc}
\alias{diff_d10a}
\alias{diff_d10i}
\alias{diff_d50a}
\alias{diff_d50i}
\alias{diff_d90a}
\alias{diff_d90i}
\alias{diff_dc}
\alias{diff_ed}
\alias{diff_ew}
\alias{diff_f}
\alias{diff_fnc}
\alias{diff_ic}
\alias{diff_ie}
\alias{diff_meic}
\alias{diff_meie}
\alias{diff_mnc}
\alias{diff_s}
\alias{diff_sc}
\alias{diff_ss}
\alias{diff_table}
\alias{diff_v}
\alias{diff_ws}
\title{Difference Table}
\usage{
diff_dc(d1, d2)

diff_ed(d1, d2)

diff_ew(d1, d2)

diff_f(d1, d2)

diff_fnc(d1, d2)

diff_ic(d1, d2)

diff_ie(d1, d2)

diff_mnc(d1, d2)

diff_meic(d1, d2)

diff_meie(d1, d2)

diff_sc(d1, d2)

diff_ss(d1, d2)

diff_s(d1, d2)

diff_v(d1, d2)

diff_ws(d1, d2)

diff_d10a(d1, d2)

diff_d10i(d1, d2)

diff_d50a(d1, d2)

diff_d50i(d1, d2)

diff_d90a(d1, d2)

diff_d90i(d1, d2)

diff_table(d1, d2, tcol, diffcol)
}
\arguments{
\item{d1}{The first dataframe.}

\item{d2}{The second dataframe.}

\item{tcol}{The time column name.}

\item{diffcol}{The name of the difference column to be created.}
}
\value{
A dataframe.
}
\description{
Compute a difference table.
}
\section{Functions}{
\itemize{
\item \code{diff_dc}: Compute a difference table for Dredged Cumulative data.

\item \code{diff_ed}: Compute a difference table for Effective Depth data.

\item \code{diff_ew}: Compute a difference table for Effective Width data.

\item \code{diff_f}: Compute a difference table for Flow data.

\item \code{diff_fnc}: Compute a difference table for Froude Number Channel data.

\item \code{diff_ic}: Compute a difference table for Invert Change data.

\item \code{diff_ie}: Compute a difference table for Invert Elevation data.

\item \code{diff_mnc}: Compute a difference table for Manning's n Channel data.

\item \code{diff_meic}: Compute a difference table for Mean Effective Invert Change data.

\item \code{diff_meie}: Compute a difference table for Mean Effective Invert Elevation data.

\item \code{diff_sc}: Compute a difference table for Sediment Concentration data.

\item \code{diff_ss}: Compute a difference table for Shear Stress data.

\item \code{diff_s}: Compute a difference table for Slope data.

\item \code{diff_v}: Compute a difference table for Velocity data.

\item \code{diff_ws}: Compute a difference table for Water Surface data.

\item \code{diff_d10a}: Compute a difference table for d10 Active data.

\item \code{diff_d10i}: Compute a difference table for d10 Inactive data.

\item \code{diff_d50a}: Compute a difference table for d50 Active data.

\item \code{diff_d50i}: Compute a difference table for d50 Inactive data.

\item \code{diff_d90a}: Compute a difference table for d90 Active data.

\item \code{diff_d90i}: Compute a difference table for d90 Inactive data.
}}

