% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{diff_wse}
\alias{diff_wse}
\title{Water Surface Elevation Difference Table}
\usage{
diff_wse(d1, d2, tcol = "Time")
}
\arguments{
\item{d1}{The first dataframe.}

\item{d2}{The second dataframe.}

\item{tcol}{the time column name.}
}
\value{
A dataframe.
}
\description{
Compute a difference table for water surface elevation data.
}

