% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{difference_table}
\alias{difference_table}
\alias{difference_dredged_cum}
\alias{difference_effective_depth}
\alias{difference_effective_width}
\alias{difference_flow}
\alias{difference_froude_number_channel}
\alias{difference_hydraulic_radius}
\alias{difference_invert_change}
\alias{difference_invert_elevation}
\alias{difference_mannings_n_channel}
\alias{difference_mean_effective_invert_change}
\alias{difference_mean_effective_invert_elevation}
\alias{difference_moveable_elv_l}
\alias{difference_moveable_elv_r}
\alias{difference_moveable_sta_l}
\alias{difference_moveable_sta_r}
\alias{difference_observed_data}
\alias{difference_sediment_concentration}
\alias{difference_shear_stress}
\alias{difference_shear_velocity}
\alias{difference_slope}
\alias{difference_temperature}
\alias{difference_thickness_cover}
\alias{difference_thickness_inactive}
\alias{difference_thickness_subsurface}
\alias{difference_velocity}
\alias{difference_water_surface}
\alias{difference_d10_active}
\alias{difference_d10_cover}
\alias{difference_d10_inactive}
\alias{difference_d10_subsurface}
\alias{difference_d16_active}
\alias{difference_d16_cover}
\alias{difference_d16_inactive}
\alias{difference_d16_subsurface}
\alias{difference_d50_active}
\alias{difference_d50_cover}
\alias{difference_d50_inactive}
\alias{difference_d50_subsurface}
\alias{difference_d84_active}
\alias{difference_d84_cover}
\alias{difference_d84_inactive}
\alias{difference_d84_subsurface}
\alias{difference_d90_active}
\alias{difference_d90_cover}
\alias{difference_d90_inactive}
\alias{difference_d90_subsurface}
\title{Difference Table}
\usage{
difference_table(d1, d2, difference.col = "Difference", relative = FALSE,
  time.col = "Time")

difference_dredged_cum(d1, d2, relative = FALSE)

difference_effective_depth(d1, d2, relative = FALSE)

difference_effective_width(d1, d2, relative = FALSE)

difference_flow(d1, d2, relative = FALSE)

difference_froude_number_channel(d1, d2, relative = FALSE)

difference_hydraulic_radius(d1, d2, relative = FALSE)

difference_invert_change(d1, d2, relative = FALSE)

difference_invert_elevation(d1, d2, relative = FALSE)

difference_mannings_n_channel(d1, d2, relative = FALSE)

difference_mean_effective_invert_change(d1, d2, relative = FALSE)

difference_mean_effective_invert_elevation(d1, d2, relative = FALSE)

difference_moveable_elv_l(d1, d2, relative = FALSE)

difference_moveable_elv_r(d1, d2, relative = FALSE)

difference_moveable_sta_l(d1, d2, relative = FALSE)

difference_moveable_sta_r(d1, d2, relative = FALSE)

difference_observed_data(d1, d2, relative = FALSE)

difference_sediment_concentration(d1, d2, relative = FALSE)

difference_shear_stress(d1, d2, relative = FALSE)

difference_shear_velocity(d1, d2, relative = FALSE)

difference_slope(d1, d2, relative = FALSE)

difference_temperature(d1, d2, relative = FALSE)

difference_thickness_cover(d1, d2, relative = FALSE)

difference_thickness_inactive(d1, d2, relative = FALSE)

difference_thickness_subsurface(d1, d2, relative = FALSE)

difference_velocity(d1, d2, relative = FALSE)

difference_water_surface(d1, d2, relative = FALSE)

difference_d10_active(d1, d2, relative = FALSE)

difference_d10_cover(d1, d2, relative = FALSE)

difference_d10_inactive(d1, d2, relative = FALSE)

difference_d10_subsurface(d1, d2, relative = FALSE)

difference_d16_active(d1, d2, relative = FALSE)

difference_d16_cover(d1, d2, relative = FALSE)

difference_d16_inactive(d1, d2, relative = FALSE)

difference_d16_subsurface(d1, d2, relative = FALSE)

difference_d50_active(d1, d2, relative = FALSE)

difference_d50_cover(d1, d2, relative = FALSE)

difference_d50_inactive(d1, d2, relative = FALSE)

difference_d50_subsurface(d1, d2, relative = FALSE)

difference_d84_active(d1, d2, relative = FALSE)

difference_d84_cover(d1, d2, relative = FALSE)

difference_d84_inactive(d1, d2, relative = FALSE)

difference_d84_subsurface(d1, d2, relative = FALSE)

difference_d90_active(d1, d2, relative = FALSE)

difference_d90_cover(d1, d2, relative = FALSE)

difference_d90_inactive(d1, d2, relative = FALSE)

difference_d90_subsurface(d1, d2, relative = FALSE)
}
\arguments{
\item{d1}{The first dataframe, considered the "base" result.}

\item{d2}{The second dataframe, considered the "new" result.}

\item{difference.col}{The name of the difference column to be created.}

\item{relative}{Logical: report differences as relative difference.}

\item{time.col}{The time column name.}
}
\value{
A dataframe, with difference defined as \code{d2- d1}.
  if \code{relative = TRUE}, the difference is defined as
  \code{(d2 - d1)/(0.5*(d2 + d1))}.
}
\description{
Compute a difference table.
}
\section{Functions}{
\itemize{
\item \code{difference_dredged_cum}: Compute a difference table for Dredged Cum data.

\item \code{difference_effective_depth}: Compute a difference table for Effective Depth data.

\item \code{difference_effective_width}: Compute a difference table for Effective_Width data.

\item \code{difference_flow}: Compute a difference table for Flow data.

\item \code{difference_froude_number_channel}: Compute a difference table for Froude Number Channel data.

\item \code{difference_hydraulic_radius}: Compute a difference table for Hydraulic Radius data.

\item \code{difference_invert_change}: Compute a difference table for Invert Change data.

\item \code{difference_invert_elevation}: Compute a difference table for Invert Elevation data.

\item \code{difference_mannings_n_channel}: Compute a difference table for Mannings n Channel data.

\item \code{difference_mean_effective_invert_change}: Compute a difference table for Mean Effective Invert Change data.

\item \code{difference_mean_effective_invert_elevation}: Compute a difference table for Mean Effective Invert Elevation data.

\item \code{difference_moveable_elv_l}: Compute a difference table for Moveable Elv L data.

\item \code{difference_moveable_elv_r}: Compute a difference table for Moveable Elv R data.

\item \code{difference_moveable_sta_l}: Compute a difference table for Moveable Sta L data.

\item \code{difference_moveable_sta_r}: Compute a difference table for Moveable Sta R data.

\item \code{difference_observed_data}: Compute a difference table for Observed Data data.

\item \code{difference_sediment_concentration}: Compute a difference table for Sediment Concentration data.

\item \code{difference_shear_stress}: Compute a difference table for Shear Stress data.

\item \code{difference_shear_velocity}: Compute a difference table for Shear Velocity data.

\item \code{difference_slope}: Compute a difference table for Slope data.

\item \code{difference_temperature}: Compute a difference table for Temperature data.

\item \code{difference_thickness_cover}: Compute a difference table for Thickness Cover data.

\item \code{difference_thickness_inactive}: Compute a difference table for Thickness Inactive data.

\item \code{difference_thickness_subsurface}: Compute a difference table for Thickness Subsurface data.

\item \code{difference_velocity}: Compute a difference table for Velocity data.

\item \code{difference_water_surface}: Compute a difference table for Water Surface data.

\item \code{difference_d10_active}: Compute a difference table for d10 Active data.

\item \code{difference_d10_cover}: Compute a difference table for d10 Cover data.

\item \code{difference_d10_inactive}: Compute a difference table for d10 Inactive data.

\item \code{difference_d10_subsurface}: Compute a difference table for d10 Subsurface data.

\item \code{difference_d16_active}: Compute a difference table for d16 Active data.

\item \code{difference_d16_cover}: Compute a difference table for d16 Cover data.

\item \code{difference_d16_inactive}: Compute a difference table for d16 Inactive data.

\item \code{difference_d16_subsurface}: Compute a difference table for d16 Subsurface data.

\item \code{difference_d50_active}: Compute a difference table for d50 Active data.

\item \code{difference_d50_cover}: Compute a difference table for d50 Cover data.

\item \code{difference_d50_inactive}: Compute a difference table for d50 Inactive data.

\item \code{difference_d50_subsurface}: Compute a difference table for d50 Subsurface data.

\item \code{difference_d84_active}: Compute a difference table for d84 Active data.

\item \code{difference_d84_cover}: Compute a difference table for d84 Cover data.

\item \code{difference_d84_inactive}: Compute a difference table for d84 Inactive data.

\item \code{difference_d84_subsurface}: Compute a difference table for d84 Subsurface data.

\item \code{difference_d90_active}: Compute a difference table for d90 Active data.

\item \code{difference_d90_cover}: Compute a difference table for d90 Cover data.

\item \code{difference_d90_inactive}: Compute a difference table for d90 Inactive data.

\item \code{difference_d90_subsurface}: Compute a difference table for d90 Subsurface data.
}}

\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.flow = read_standard(simple.quasi, "Flow")

difference_table(quasi.flow, quasi.flow)

quasi.double = operate_table(quasi.flow, fun = function(x) 2*x)
difference_table(quasi.flow, quasi.double)
difference_table(quasi.flow, quasi.double, relative = TRUE)

}
