% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{fix_xsviewer_stations}
\alias{fix_xsviewer_stations}
\title{Fix Cross Section Viewer Stations}
\usage{
fix_xsviewer_stations(d, station.data, station.cols = c("downstream.station",
  "upstream.station"))
}
\arguments{
\item{d}{The cross section viewer data.}

\item{station.data}{A two-column dataframe containing the actual (column 1)
and offset (column 2) river stations.}

\item{station.cols}{The names of columns containing station data to be
adjusted.}
}
\value{
The data frame \code{d} with adjusted station values.
}
\description{
Replace the cross section viewer stations with the actual stations.
}
\details{
Some versions of the cross section viewer did not allow station
  IDs above a certain value, requiring some users to offset their actual
  river stations to use the software. This function reverses the offset or
  scaling.
}
