% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{force_RAS_version}
\alias{force_RAS_version}
\title{Force RAS Version}
\usage{
force_RAS_version(version)
}
\arguments{
\item{version}{The RAS version to use.}
}
\description{
Force RAStestR to assume the specified RAS version. Useful for
working with RAS development environments or incomplete datasets 
that do not include RAS version info.
}
