% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{get_dataset}
\alias{get_dataset}
\title{Read HDF Table}
\usage{
get_dataset(x, path, type)
}
\arguments{
\item{x}{an HDF object, e.g. output of \code{h5file()}.}

\item{path}{The table path.}

\item{type}{The type of data contained in the table. Can be one of "double",
"integer", or "character".}
}
\description{
Safely read in a specific HDF table. This function is used internally and
should not be called directly by the user.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
tryCatch({
    test.table = "Geometry/Cross Sections/Bank Stations"
    quasi.h5 = h5::h5file(simple.quasi)
    RAStestR:::get_dataset(quasi.h5, test.table, "double")
  }, finally = {
    h5::h5close(quasi.h5)
  }) 

}
