% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{get_group_attr}
\alias{get_group_attr}
\title{Read HDF Attributes}
\usage{
get_group_attr(x, path = NULL, attrs = NULL)
}
\arguments{
\item{x}{an HDF object, e.g. output of \code{h5file()}.}

\item{path}{The table path.}

\item{attrs}{The attributes to read. If \code{NULL}, all 
attributes will be read.}
}
\value{
a named list of attributes.
}
\description{
Safely read attributes of an HDF Group
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")

tryCatch({
    quasi.h5 = h5::h5file(simple.quasi)
    RAStestR:::get_group_attr(quasi.h5)
  }, finally = {
    h5::h5close(quasi.h5)
  }) 

tryCatch({
    test.group = "Plan Data/Plan Information"
    quasi.h5 = h5::h5file(simple.quasi)
    RAStestR:::get_group_attr(quasi.h5, test.group)
  }, finally = {
    h5::h5close(quasi.h5)
  }) 

}
