% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{get_meta}
\alias{get_meta}
\title{Get RAS Plan Meta Data}
\usage{
get_meta(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
A named list of plan meta data.
}
\description{
Get meta data of RAS plan.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::get_meta(simple.quasi)

}
