% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{get_RAS_version}
\alias{get_RAS_version}
\title{Get RAS Version}
\usage{
get_RAS_version(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
The RAS version.
}
\description{
Get the RAS version that generated the specified file.
}
