% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{get_run_type}
\alias{get_run_type}
\title{Get RAS Plan Run Type}
\usage{
get_run_type(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\description{
Identify a RAS plan as Unsteady, Steady or QuasiUnsteady.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::get_run_type(simple.quasi)

simple.unsteady= system.file("sample-data/SampleUnsteady.hdf",
  package = "RAStestR")
RAStestR:::get_run_type(simple.unsteady)

}
