% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_2d_coordinates}
\alias{list_2d_coordinates}
\title{List 2D Area Coordinates}
\usage{
list_2d_coordinates(f, which.area, type = c("FacePoint", "Cell"))
}
\arguments{
\item{f}{The HDF5 file to read.}

\item{which.area}{The 2D area to get the coordinates for.}

\item{type}{The type of coordinates to get, i.e. `FacePoint`
or `Cell` (center) coordinates.}
}
\value{
a dataframe of XY coordinates for the 2D area.
}
\description{
list the face point or cell center coordinates of a 2D area.
}
