% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_2dareas}
\alias{list_2dareas}
\title{List 2D Areas}
\usage{
list_2dareas(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a vector of 2D flow area names.
}
\description{
List 2D flow areas
}
