% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_bank_stations}
\alias{list_bank_stations}
\title{List Bank Stations}
\usage{
list_bank_stations(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a vector of bank stations
}
\description{
List bank stations.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_bank_stations(simple.quasi)

}
