% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_grain_classes}
\alias{list_grain_classes}
\title{List Grain Classes}
\usage{
list_grain_classes(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a vector of grain glass labels.
}
\description{
Get list of RAS sediment grain class labels.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_grain_classes(simple.quasi)

}
