% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_grains}
\alias{list_grains}
\title{List Grain Classes}
\usage{
list_grains(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a vector of grain glass labels.
}
\description{
Get list of RAS sediment grain class labels.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_grains(simple.quasi)

}
