% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_lengths}
\alias{list_lengths}
\title{List Cross Section Lengths}
\usage{
list_lengths(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a vector of cross section lengths.
}
\description{
Get list cross section lengths.
}
