% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_levees}
\alias{list_levees}
\title{List Levees}
\usage{
list_levees(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a matrix of levee stations and elevations. Column order is:
  left station, left elevation, right station, right elevation.
}
\description{
List levee stations and elevations.
}
