% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_node_descriptions}
\alias{list_node_descriptions}
\title{List Node Descriptions}
\usage{
list_node_descriptions(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a vector of node descriptions.
}
\description{
List node descriptions.
}
