% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_node_names}
\alias{list_node_names}
\title{List Node Names}
\usage{
list_node_names(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a vector of node names.
}
\description{
List node names.
}
