% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_output_times}
\alias{list_output_times}
\title{List Output Times}
\usage{
list_output_times(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a vector of cross section lengths.
}
\description{
List output times.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_output_times(simple.quasi)

}
