% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_plan_info}
\alias{list_plan_info}
\title{List Plan Information}
\usage{
list_plan_info(f, print = TRUE)
}
\arguments{
\item{f}{The hDF5 file to read.}

\item{print}{Print the plan information to the console in a nice format.}
}
\value{
A list of plan information.
}
\description{
List Basic HEC-RAS plan information.
}
