% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_reaches}
\alias{list_reaches}
\title{List Reach Names}
\usage{
list_reaches(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a vector of reach names.
}
\description{
List RAS reach names.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_reaches(simple.quasi)

}
