% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_rivers}
\alias{list_rivers}
\title{List River Names}
\usage{
list_rivers(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a vector of river names.
}
\description{
List RAS river names.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_rivers(simple.quasi)

}
