% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_stations}
\alias{list_stations}
\title{Read River Station Table}
\usage{
list_stations(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a vector of river station labels.
}
\description{
Read RAS river station labels.
}
