% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{list_tables}
\alias{list_tables}
\title{HEC-RAS Output Tables}
\usage{
list_tables()
}
\value{
A list named list of character vectors containing the names
  of "standard" and "sediment" tables.
}
\description{
List all potential output tables produced by an HEC-RAS sediment
transport model.
}

