% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_xs}
\alias{list_xs}
\title{List Cross Section Tables}
\usage{
list_xs(f, xs.block)
}
\arguments{
\item{f}{The HDF5 file to read.}

\item{xs.block}{The HDF folder containing the cross section output tables.}
}
\value{
a vector of cross section output labels.
}
\description{
List cross section tables.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_xs(simple.quasi, file.path("Results", "Sediment",
    "Output Blocks", "Sediment SE", "Sediment Time Series", 
    "Cross Section SE"))

}
