% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{offset_xsviewer_long_change}
\alias{offset_xsviewer_long_change}
\title{Offset Cross Section Viewer Longitudinal Change}
\usage{
offset_xsviewer_long_change(d, offset.station,
  upstream.station.col = "upstream.station",
  downstream.station.col = "downstream.station",
  data.col = "cumulative.volume.change")
}
\arguments{
\item{d}{The cross section viewer data.}

\item{offset.station}{The station to offset the data from.}

\item{upstream.station.col}{The name of the column containing the station at
the upstream face of the control volume.}

\item{downstream.station.col}{The name of the column containing the station
at the downstream face of the control volume.}

\item{data.col}{The name of the column containing the longitudinal change
values.}
}
\value{
A subset of the data frame \code{d} containing only stations below
  \code{offset.station} and with adjusted longitudinal change values.
}
\description{
Extract a segment of a Cross Section Viewer longitudinal change curve,
offsetting data values where appropriate.
}
