% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{operate_sediment}
\alias{operate_sediment}
\title{Table Operations (Sediment)}
\usage{
operate_sediment(..., fun = "+", time.col = "Time",
  grain.col = "GrainClass")
}
\arguments{
\item{...}{Arbitrary number of wide-format data tables to combine.}

\item{fun}{A function to apply.}

\item{time.col}{The time column name.}

\item{grain.col}{the grain class column name.}
}
\value{
A single sediment table.
}
\description{
Combine sediment tables via an operation, e.g. addition or multiplication.
}
