% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{override_RAS_version}
\alias{override_RAS_version}
\title{Override RAS Version}
\usage{
override_RAS_version(v, override.v)
}
\arguments{
\item{v}{The RAS version to override.}

\item{override.v}{The RAS version to be assumed.}
}
\description{
Specify RAS version overrides. Useful when working with files produced 
from both RAS development and release versions, e.g. by assuming that
files produced with RAS development x.x are structured in the same way
as a specific release version.
}
