% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\docType{package}
\name{RAStestR-package}
\alias{RAStestR}
\alias{RAStestR-package}
\title{Automated Testing of HEC-RAS}
\description{
This package is designed to provide an automated testing environment
of HEC-RAS sediment transport model outputs. Functions are provided 
for reading, analyzing and visualizing HDF5 output data and 
generating test reports. See the vignette to get started.
}

