% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2da.r
\name{read_2d_standard}
\alias{read_2d_standard}
\title{Read 2D Area Tables}
\usage{
read_2d_standard(f, table.name, which.times = NULL, which.areas = NULL,
  override.sediment = FALSE)
}
\arguments{
\item{f}{The HDF5 file to read.}

\item{table.name}{The table to read.}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.areas}{2D flow areas to extract. Can accept either numeric
IDs or 2D flow area names. If NULL, all 2D flow areas will be 
returned.}

\item{override.sediment}{(For Unsteady+Sediment models only) If True,
extract data from the hydraulic rather than sediment output.}
}
\value{
A named list of dataframes. List names correspond to the 2D 
 flow area names. Dataframes contain a column "Time" containing the 
 Date Time Stamp data and columns "CELL_####" where ### is the 2D cell ID.
}
\description{
Read standard (not by grain class) tables for one or more 2D areas.
}
