% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{read_bed_limits}
\alias{read_bed_limits}
\title{Read Bed Limits}
\usage{
read_bed_limits(f, which.stations = NULL, which.times = NULL)
}
\arguments{
\item{f}{The HDF5 file to read.}

\item{which.stations}{Character vector of stations to extract. If
NULL, all stations will be returned.}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}
}
\value{
a dataframe with columns "Time", "Station", "LOB" and "ROB", where
  columns "LOB" and "ROB" list the left and right moveable bed stations,
  respectively.
}
\description{
Read the moveable bed limit stations.
}
\details{
The moveable bed limits will change over time in models that use
  BSTEM. In general, it is recommended but not required that the BSTEM toe
  stations are coincident with the moveable bed limits; otherwise, cross
  sections may adjust in unexpected ways.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")

read_bed_limits(simple.quasi)
read_bed_limits(simple.quasi, which.times = "11DEC1990 01:00:00",
  which.stations = "XS_800")
read_bed_limits(simple.quasi, which.times = 1:4,
  which.stations = 1:2)

}
