% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_grains}
\alias{read_grains}
\title{Read Grain Class Table}
\usage{
read_grains(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a vector of grain glass labels.
}
\description{
Read RAS sediment grain class labels.
}

