% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_hdtable}
\alias{read_hdtable}
\title{Read RAS Table}
\usage{
read_hdtable(f, table.path, row.table.path, col.table.path, run.type,
  rowcolname, colprefix)
}
\arguments{
\item{f}{The h5 file to read.}

\item{table.path}{The table to read in.}

\item{row.table.path}{The table containing the row identifiers.}

\item{col.table.path}{The table containing the column identifiers.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}

\item{rowcolname}{The name to assign to the new row id column.}

\item{colprefix}{A prefix to apply to the column IDs.}
}
\value{
A dataframe.
}
\description{
Read RAS sediment data output. This function is used internally and should
not be called directly by the user.
}
