% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{read_sediment}
\alias{read_lcmc}
\alias{read_lsmd}
\alias{read_mbcc}
\alias{read_mic}
\alias{read_moc}
\alias{read_sediment}
\title{Sediment By Grain Class Table}
\usage{
read_sediment(f, tablelab, run.type, which.grains = "", which.rows = NULL)

read_lsmd(f, run.type, which.grains, which.rows)

read_lcmc(f, run.type, which.grains, which.rows)

read_mbcc(f, run.type, which.grains, which.rows)

read_mic(f, run.type, which.grains, which.rows)

read_moc(f, run.type, which.grains, which.rows)
}
\arguments{
\item{f}{The h5 file to read.}

\item{tablelab}{The table to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}

\item{which.grains}{Grain class tables to extract. "" Corresponds to 
the totals, "1" is the first grain class, etc.}

\item{which.rows}{A numeric identifying the row numbers to extract, 
0 being the first row.}
}
\value{
A dataframe with a column "Time" containing the Date Time 
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read the sediment data output for all grain classes.
}
\section{Functions}{
\itemize{
\item \code{read_lsmd}: Read the Lateral Structure Mass Divergence data output.

\item \code{read_lcmc}: Read the Longitudinal Cumulative Mass Change data output.

\item \code{read_mbcc}: Read the Mass Bed Change Cumulative data output.

\item \code{read_mic}: Read the Mass In Cumulative data output.

\item \code{read_moc}: Read the Mass Out Cumulative data output.
}}

