% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_standard}
\alias{read_standard}
\title{Read Standard Table}
\usage{
read_standard(f, table.name, which.times = NULL, which.stations = NULL,
  override.sediment = FALSE)
}
\arguments{
\item{f}{The HDF5 file to read.}

\item{table.name}{The table to read.}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of stations to extract. If
NULL, all stations will be returned.}

\item{override.sediment}{(For Unsteady+Sediment models only) If True,
extract data from the hydraulic rather than sediment output.}
}
\value{
A dataframe with a column "Time" containing the Date Time
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read a standard (not grain class-specific) table.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
read_standard(simple.quasi, "Flow")

simple.unsteady = system.file("sample-data/SampleUnsteady.hdf",
  package = "RAStestR")
read_standard(simple.unsteady, "Flow")

read_standard(simple.quasi, "Flow", which.times = "11DEC1990 01:00:00",
  which.stations = c("XS_800", "XS_796.00*"))

read_standard(simple.quasi, "Flow", which.times = 2:3,
  which.stations = 1:4)

}
