% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{read_station_lengths}
\alias{read_station_lengths}
\title{Read River Station Lengths}
\usage{
read_station_lengths(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a dataframe with columns "Station", "LOB", "Channel" and "ROB", where
  column "Channel" list the channel length and columns "LOB" and "ROB" list
  the left and right bank stations lengths, respectively.
}
\description{
Read the river station lengths.
}
