% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{read_xs}
\alias{read_xs}
\title{Cross Section Table}
\usage{
read_xs(f, run.type, which.times = NULL, which.stations = NULL)
}
\arguments{
\item{f}{The h5 file to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of stations to extract. If
NULL, all stations will be returned.}
}
\value{
A dataframe with a column "Time" containing the Date Time
  Stamp data; column "Station" containing the station ID in format "XS_####"
  where ### is the cross-section ID; column "Distance" containing the lateral
  distance location; and column "Elevation" containing the elevation at the
  distance location.
}
\description{
Read the cross section data output.
}
