% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{read_xsviewer}
\alias{read_xsviewer}
\title{Read Cross Section Viewer Outputs}
\usage{
read_xsviewer(f, col.spec = NULL, viewer.version = "1.1.56.0")
}
\arguments{
\item{f}{The cross section viewer file to read. Typically a \code{.txt} file.}

\item{col.spec}{Define column types. If \code{NULL}, RAStestR will try to
determine the column types automatically.}

\item{viewer.version}{The version of the cross viewer used.}
}
\description{
Read data files exported from the cross section viewer.
}
