% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{read_xsviewer_long_change}
\alias{read_xsviewer_long_change}
\title{Read Cross Section Viewer Longitudinal Change}
\usage{
read_xsviewer_long_change(f, cumulative = TRUE, viewer.version = "1.1.56.0")
}
\arguments{
\item{f}{The cross section viewer file to read. Typically a \code{.txt} file.}

\item{cumulative}{If \code{TRUE}, return the cumulative volume change.
Otherwise, return the incremental control volume change.}

\item{viewer.version}{The version of the cross viewer used.}
}
\value{
a dataframe.
}
\description{
Read the longitudinal change output exported from the Cross Section Viewer.
}
