% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{readlcmc}
\alias{readlcmc}
\title{Longitudinal Cumulative Mass Change Table}
\usage{
readlcmc(f, run.type, which.grains, which.rows)
}
\arguments{
\item{run.type}{If TRUE, the data is from an unsteady flow plan
and needs some additional cleanup. Use FALSE for quasi-unsteady 
data.}

\item{which.grains}{Which grain class tables to extract data from. If
missing, all grain class tables will be extracted. Character value
"" extracts the total table, "1" the first grain class, etc.}

\item{which.rows}{Which rows to extract from each grain class table. 
0 refers to the first row of the HDF5 table. Note that this index 
refers to the cleaned-up data, so check the row ID to ensure the 
correct rows are extracted.}

\item{x}{The h5 object to read.}
}
\value{
A dataframe with a column "Time" containing the Date Time 
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read the Longitudinal Cumulative Mass Change sediment data output.
}

