% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{readsediment}
\alias{readsediment}
\title{Sediment By Grain Class Table}
\usage{
readsediment(f, tablelab, run.type, which.grains = "", which.rows = NULL)
}
\arguments{
\item{f}{The h5 object to read.}

\item{run.type}{If TRUE, the data is from an unsteady flow plan
and needs some additional cleanup. Use FALSE for quasi-unsteady 
data.}

\item{which.grains}{Grain class tables to extract. "" Corresponds to 
the totals, "1" is the first grain class, etc.}

\item{which.rows}{A numeric identifying the row numbers to extract, 
0 being the first row.}
}
\value{
A dataframe with a column "Time" containing the Date Time 
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read the sediment data output for all grain classes.
}

