% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{readwse}
\alias{readwse}
\title{Water Surface Elevation Table}
\usage{
readwse(f, run.type)
}
\arguments{
\item{f}{The h5 object to read.}

\item{run.type}{If TRUE, the data is from an unsteady flow plan
and needs some additional cleanup. Use FALSE for quasi-unsteady 
data.}
}
\value{
A dataframe with a column "Time" containing the Date Time 
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read the Water Surface Elevation sediment data output.
}

