% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{reformat_fields}
\alias{reformat_fields}
\title{Reformat Standard Fields}
\usage{
reformat_fields(ld, fields = c("Time", "Station"))
}
\arguments{
\item{ld}{A data table.}

\item{fields}{A list of fields to reformat.}
}
\value{
The data table with reformatted fields.
}
\description{
Reformat standard RAStestR fields as standard R data types.
}
\details{
If the "Time" field is specified, values in that column
  will be formatted as R timestamps (POSIXct). If the "Station"
  field is specified, stations will be converted to numeric values
  (only applicable to long format tables).
}
