% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{region_to_volume}
\alias{region_to_volume}
\title{Region Area to Volume}
\usage{
region_to_volume(d, time.col = "Time", region.col = "Region",
  station.lengths = NULL)
}
\arguments{
\item{d}{The area data, e.g. output of \code{xs_area}.}

\item{time.col}{The time column name.}

\item{region.col}{The name of the column containing the Region identifier.}

\item{station.lengths}{The cross section length values.}
}
\value{
A wide-format table of cross section region volumes.
}
\description{
Compute cross section volumes by region.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")

quasi.xs = read_xs(simple.quasi)
quasi.banks = read_bank_stations(simple.quasi)
quasi.extents = xs_extents(quasi.xs)
quasi.lengths = xs_lengths(read_station_lengths(simple.quasi))

quasi.regions = xs_regions(quasi.xs, bank.stations = quasi.banks, 
 extent.stations = quasi.extents)
region_to_volume(quasi.regions, station.lengths = quasi.lengths)

}
