% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{rename_interpolated_xs}
\alias{rename_interpolated_xs}
\title{Rename Interpolated Cross Sections}
\usage{
rename_interpolated_xs(d)
}
\arguments{
\item{d}{A data table containing interpolated cross section data.}
}
\value{
The data frame \code{d} with reformatted identifiers for interpolated
  cross sections.
}
\description{
Rename the identifiers of interpolated cross sections.
}
\details{
Interpolated cross sections are identified by the presence of a
  '*' in the column name or value of the "Station" column (for long-format
  data). The '*' symbol can interfere with certain selections or data
  manipulations. This function removes the '*' symbol from the Station
  identifiers.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.flow = read_standard(simple.quasi, "Flow")
rename_interpolated_xs(quasi.flow)

}
