% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{rmse_table}
\alias{rmse_ic}
\alias{rmse_table}
\alias{rmse_ws}
\title{Root Mean Square Error Table}
\usage{
rmse_table(d, groupcol, diffcol, rmsecol)

rmse_ic(d, groupcol = "Station")

rmse_ws(d, groupcol = "Station")
}
\arguments{
\item{d}{The difference table.}

\item{groupcol}{the column to group differences by.}

\item{diffcol}{the column containing difference values.}

\item{rmsecol}{The output column containing RMSE values}
}
\value{
A dataframe.
}
\description{
Compute RMSE from a difference table.
}
\section{Functions}{
\itemize{
\item \code{rmse_ic}: Compute RMSE of Invert Change outputs.

\item \code{rmse_ws}: Compute RMSE of Water Surface outputs.
}}

