% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{station_to_distance}
\alias{station_to_distance}
\title{Convert Stations To Distance}
\usage{
station_to_distance(ld, distance.col, station.col = "station",
  metric = FALSE)
}
\arguments{
\item{ld}{The long-format data, i.e. output from 
\code{to_longtable}.}

\item{distance.col}{The new column to hold distance values.}

\item{station.col}{The column containing station values.}

\item{metric}{If \code{TRUE}, stations are assumed to be in 
kilometers and distances are returned in meters. If \code{FALSE}, 
stations are assumed to be in miles and distances are returned in 
feet.}
}
\value{
The data table with an additional column of distances.
}
\description{
Convert station values to distance upstream.
}

