% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_change}
\alias{survey_change}
\title{Survey Longitudinal Change}
\usage{
survey_change(d, station.lengths)
}
\arguments{
\item{d}{The area data, e.g. output of \code{xs_area}.}

\item{station.lengths}{The cross section length values.}
}
\description{
Compute longitudinal cumulative change from survey data.
}
\details{
This function performs similarly to \code{cumulative_table} but is
  designed to work with survey data rather than RAS outputs. The primary
  difference is that this function allows the computation of longitudinal
  cumulative change when some surveys are incomplete. The function recomputes
  the station lengths for each survey based on what cross sections are
  present in the survey. The longitudinal cumulative volume is then computed
  with missing stations omitted from the data. The longitudinal volume curves
  are then linearly interpolated across the missing stations prior to
  computing longitudinal change.
}
