% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_clip}
\alias{survey_clip}
\title{Clip Survey Cross Section}
\usage{
survey_clip(d, extent.stations, station.col = "Station", time.col = "Time",
  distance.col = "Distance", elevation.col = "Elevation")
}
\arguments{
\item{d}{The survey data in cross section format.}

\item{extent.stations}{The maximum extents of each station. Expects a
three-column table with column names "Station", "LE" and "RE, where "LE"
and "RE" define the left extent and right extent of each station,
respectively. Can also accept a two-element numeric vector specifying
constant left and right edges. If NULL, the minimum and maximum coincident
distances of each cross section will be used.}

\item{station.col}{The column containing cross section station IDs.}

\item{time.col}{The column containing time stamps.}

\item{distance.col}{The column containing cross section distances.}

\item{elevation.col}{The column containing cross section elevations.}
}
\value{
The clipped cross section data.
}
\description{
Clip survey cross sections to a set of specified extent stations.
}
