% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_extend}
\alias{survey_extend}
\title{Extend Survey Cross Section}
\usage{
survey_extend(d, station.col = "Station", time.col = "Time",
  distance.col = "Distance", mode = c("lag", "lead"))
}
\arguments{
\item{d}{The survey data in cross section format.}

\item{station.col}{The column containing cross section station IDs.}

\item{time.col}{The column containing time stamps.}

\item{distance.col}{The column containing cross section distances.}

\item{mode}{Extend cross sections by using data from prior cross section
(lag) or future cross section (lead).}
}
\value{
The extended cross section data.
}
\description{
Extend survey cross sections to the specified extents.
}
