% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_interp}
\alias{survey_interp}
\title{Interpolate Survey Stations}
\usage{
survey_interp(d, drop.stations = FALSE)
}
\arguments{
\item{d}{A wide-format table of processed data, e.g. output of
\code{xs_area}.}

\item{drop.stations}{Logical: Drop stations that cannot be interpolated.}
}
\value{
The data frame \code{d} with some or all \code{NA} values filled in.
}
\description{
Interpolate data over stations. Useful for working with surveys that have
partial overlap.
}
