% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_substitute}
\alias{survey_substitute}
\title{Substitute Survey}
\usage{
survey_substitute(d, substitute.stations, mode = c("lag", "lead"),
  station.col = "Station", time.col = "Time", distance.col = "Distance",
  elevation.col = "Elevation")
}
\arguments{
\item{d}{The survey data in cross section format.}

\item{substitute.stations}{The stations to substitute}

\item{mode}{Extend cross sections by using data from prior cross section
(lag) or future cross section (lead).}

\item{station.col}{The column containing cross section station IDs.}

\item{time.col}{The column containing time stamps.}

\item{distance.col}{The column containing cross section distances.}

\item{elevation.col}{The column containing cross section elevations.}
}
\value{
The cross section data with substituted regions
}
\description{
Substitute a portion of a cross section.
}
