% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_to_geometry}
\alias{survey_to_geometry}
\title{Cross Section Viewer To RAS Geometry}
\usage{
survey_to_geometry(f, river, reach, out.format = c("XYZ", "SE"),
  which.survey = NULL, viewer.version = "1.1.56.0")
}
\arguments{
\item{f}{The cross section viewer file to read. The function expects that
the file was exported in standard csv format, NOT as HEC format.}

\item{river}{The river label to assign to the survey data.}

\item{reach}{The reach label to assign to the survey data.}

\item{out.format}{The output format to use. If \code{out.format = "XYZ"}, the
geometry is output in X, Y, Z format (requires that \code{f} contains data
in the "Northing" and "Easting" columns). If \code{out.format = "SE"}
the geometry is output in Station-Elevation format.}

\item{which.survey}{The name of the survey to extract. Only one survey can
be converted at a time. If \code{NULL}, the function assumes that only
one survey is listed in \code{f}.}

\item{viewer.version}{The version of the cross viewer used.}
}
\value{
The survey in a format ready for import into HEC RAS.
}
\description{
Convert a csv file exported from the Cross Section Viewer to a RAS geometry
file.
}
