% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_to_xs}
\alias{survey_to_xs}
\title{Survey To Cross Section Data}
\usage{
survey_to_xs(d, viewer.version = "1.1.56.0")
}
\arguments{
\item{d}{The survey data to reformat, i.e. output of \code{read_survey}.}

\item{viewer.version}{The version of the cross viewer used.}
}
\description{
Reformat survey data to match the cross section data format of RAStestR,
e.g. the output of \code{read_xs}.
}
