% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{to_survey_template}
\alias{to_survey_template}
\title{Cross Sections To Survey Template}
\usage{
to_survey_template(d, survey.name, viewer.version = "1.1.56.0")
}
\arguments{
\item{d}{The cross section data to format.}

\item{survey.name}{The survey name. If missing, the function will search for
a column in d that matches the Cross Section Viewer template.}

\item{viewer.version}{The version of the cross viewer used.}
}
\value{
A dataframe formatted according to the template.
}
\description{
Format cross section data according to the Cross Section Viewer survey import
template and write to the clipboard.
}
