% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{to_widetable}
\alias{to_widetable}
\title{Reformat As Wide Table}
\usage{
to_widetable(d, key.col, value.col, key.prefix)
}
\arguments{
\item{d}{The data table, i.e. output from \code{to_longtable}.}

\item{key.col}{The name of the column holding data keys.}

\item{value.col}{The name of the column holding data values.}

\item{key.prefix}{Text to prepend to the new columns created from
keys in \code{key.col}.}
}
\value{
The original data table in wide format.
}
\description{
Reformat RAS data from long table format (i.e. output of
\code{to_longtable}) to wide table format (key values are mapped to
individual columns).
}
