% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{volume_to_change}
\alias{volume_to_change}
\title{Volume to Change}
\usage{
volume_to_change(d, time.col = "Time")
}
\arguments{
\item{d}{The volume data, e.g. output of \code{area_to_volume}.}

\item{time.col}{The time column name.}
}
\value{
A wide-format table of cross section volume change. A change value of
  \code{NA} is assigned to all cross section for the first time step.
}
\description{
Compute volume change over time.
}
