% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_cumulative_change}
\alias{xs_cumulative_change}
\title{Cross Section Longitudinal Cumulative Volume Change}
\usage{
xs_cumulative_change(d, station.col = "Station", time.col = "Time",
  direction = "downstream")
}
\arguments{
\item{d}{The data table of cross section volumes, e.g. output from \code{xs_volume}}

\item{station.col}{The column containing Station IDs.}

\item{time.col}{The column containing times.}

\item{direction}{Accumulate data in the downstream (descending order of cross
section IDs) or upstream (ascending order) direction. Ignored if
\code{longitudinal} is \code{FALSE}.}
}
\value{
A data frame of longitudinal cumulative volume change in wide-table
  format.
}
\description{
Build a longitudinal cumulative volume change curve from cross section data.
}
\details{
This function is simply a wrapper that applies \code{xs_volume},
  \code{xs_volume_to_change}, and \code{accumulate_data} to generate a
  longitudinal cumulative volume change curve.
}
