% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_extents}
\alias{xs_extents}
\title{Cross Section Extents}
\usage{
xs_extents(d, station.col = "Station", time.col = "Time",
  distance.col = "Distance")
}
\arguments{
\item{d}{The cross section data, i.e. output of \code{read_xs}.}

\item{station.col}{The column containing cross section station IDs.}

\item{time.col}{The column containing time stamps.}

\item{distance.col}{The column containing cross section distances.}
}
\value{
A dataframe with columns "Station", "LE" and "RE", where "LE" and
  "RE" are the left and right extents, respectively.
}
\description{
Compute the coincident extents of a cross section
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")

quasi.xs = read_xs(simple.quasi)
xs_extents (quasi.xs)

}
