% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_format}
\alias{xs_format}
\title{Format Cross Section Data}
\usage{
xs_format(d, river, reach, station.col = "Station",
  distance.col = "Distance", elevation.col = "Elevation")
}
\arguments{
\item{d}{The cross section data, i.e. output of \code{read_xs}.}

\item{river}{The River name (used by RAS).}

\item{reach}{The Reach name (used by RAS).}

\item{station.col}{The column containing cross section station IDs.}

\item{distance.col}{The column containing cross section distances.}

\item{elevation.col}{The column containing cross section elevations.}
}
\value{
A dataframe with columns formatted to match the RAS csv import format.
}
\description{
Format the cross section data for import by RAS.
}
