% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_region_cumulative_change}
\alias{xs_region_cumulative_change}
\title{Cross Section Region Cumulative Change}
\usage{
xs_region_cumulative_change(d, time.col = "Time", station.col = "Station",
  distance.col = "Distance", elevation.col = "Elevation", bank.stations,
  extent.stations = NULL, reference.elevation = NULL, region = c("LOB",
  "Channel", "ROB"), station.lengths = NULL, over.time = TRUE,
  longitudinal = TRUE, direction = "downstream")
}
\arguments{
\item{d}{The cross section data, i.e. output of \code{read_xs}.}

\item{time.col}{The column containing time stamps.}

\item{station.col}{The column containing cross section station IDs.}

\item{distance.col}{The column containing cross section distances.}

\item{elevation.col}{The column containing cross section elevations.}

\item{bank.stations}{A table defining the bank station distances to use when
computing cross section area. If \code{NULL}, the full extent of each cross
section will be used. See details for more information.}

\item{extent.stations}{The maximum extents of each station. Expects a
three-column table with column names "Station", "LE" and "RE, where "LE"
and "RE" define the left extent and right extent of each station,
respectively. Can also accept a two-element numeric vector specifying
constant left and right edges. If NULL, the minimum and maximum coincident
distances of each cross section will be used.}

\item{reference.elevation}{Use predefined reference elevations when computing
cross section area. Can be a constant or a two column station of cross
  section labels and associated reference elevation. If \code{NULL}, the
  maximum elevation of each cross section will be used.}

\item{region}{The region(s) to compute area for. Can be any combination of
"channel", "LOB", "ROB". See details for more information.}

\item{station.lengths}{The cross section length values.}

\item{over.time}{If \code{TRUE}, accumulate data across time steps. This
is generally valid only for data output at the computation time step.}

\item{longitudinal}{If \code{TRUE}, accumulate data along the reach. This
is generally only valid when all cross sections are included in \code{d}.}

\item{direction}{Accumulate data in the downstream (descending order of cross
section IDs) or upstream (ascending order) direction. Ignored if
\code{longitudinal} is \code{FALSE}.}
}
\description{
Compute cumulative cross section change by region directy from cross section
data.
}
\details{
This is essentially a wrapper for processing cross section data
  through the sequence \code{xs_regions} --> \code{region_to_volume} -->
  \code{change_sediment} --> \code{cumulative_sediment}.
}
