% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_regions}
\alias{xs_regions}
\title{Cross Section Region Area}
\usage{
xs_regions(d, time.col = "Time", station.col = "Station",
  distance.col = "Distance", elevation.col = "Elevation", bank.stations,
  extent.stations = NULL, reference.elevation = NULL,
  bottom.elevation = NULL, region = c("LOB", "Channel", "ROB"))
}
\arguments{
\item{d}{The cross section data, i.e. output of \code{read_xs}.}

\item{time.col}{The column containing time stamps.}

\item{station.col}{The column containing cross section station IDs.}

\item{distance.col}{The column containing cross section distances.}

\item{elevation.col}{The column containing cross section elevations.}

\item{bank.stations}{A table defining the bank station distances to use when
computing cross section area. If \code{NULL}, the full extent of each cross
section will be used. See details for more information.}

\item{extent.stations}{The maximum extents of each station. Expects a
three-column table with column names "Station", "LE" and "RE, where "LE"
and "RE" define the left extent and right extent of each station,
respectively. Can also accept a two-element numeric vector specifying
constant left and right edges. If NULL, the minimum and maximum coincident
distances of each cross section will be used.}

\item{reference.elevation}{Use predefined reference elevations when computing
cross section area. Can be a constant or a two column station of cross
  section labels and associated reference elevation. If \code{NULL}, the
  maximum elevation of each cross section will be used.}

\item{bottom.elevation}{Use predefined bottom elevations when computing
cross section area. Can be a constant or a two column station of cross
  section labels and associated bottom elevation. If \code{NULL}, the
  minimum elevation of each cross section will be used.}

\item{region}{The region(s) to compute area for. Can be any combination of
"channel", "LOB", "ROB". See details for more information.}
}
\value{
A wide-format table with additional column "Region" specifying the
  region for which area was computed.
}
\description{
Compute cross section area by region.
}
\details{
The \code{extent.stations} argument is similar to the
  \code{bank.stations} argument, but instead defines the coincident extent
  of each cross section. Unlike \code{bank.stations}, cross section
  extents are not allowed to vary through time.

  The \code{region} argument is used to separate cross section area
  computations based on bank and extent definitions. The "LOB" region is
  bounded by the left extent and the left bank station; the "Channel" is
  bounded by the left and right bank stations; and the "ROB" region is
  bounded by the right bank station and the right extent.
}
