% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_shift}
\alias{xs_shift}
\title{Shift Cross Sections}
\usage{
xs_shift(d, what = c("horizontal", "vertical"), shift.table,
  station.col = "Station", distance.col = "Distance",
  elevation.col = "Elevation", subtract = TRUE)
}
\arguments{
\item{d}{The cross section data.}

\item{what}{Apply a \code{"horizontal"} and/or \code{"vertical"} shift to
data.}

\item{shift.table}{A table of cross section stations and shift values.
If \code{what} is \code{"horizontal"} (\code{"vertical"}),
a two-column table is expected with the first column listing the
cross section stations and the second column listing the horizontal
(vertical) shift to apply to that cross section. If
\code{what = c("horizontal", "vertical")}, a three-column table is expected
with the second and third columns containing the horizontal and vertical
shifts, respectively.}

\item{station.col}{The column in \code{d} containing the cross section
stations.}

\item{distance.col}{The column in \code{d} containing the distance
(horizontal) values.}

\item{elevation.col}{The column in \code{d} containing the elevation
(vertical) values.}

\item{subtract}{If \code{TRUE}, values in \code{shift.table} will be
subtracted from \code{d}; if \code{FALSE}, values will be added.}
}
\value{
The data frame \code{d} with cross sections shifted according to
  \code{shift.table}
}
\description{
Apply a horizontal and/or vertical shift to cross section data.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")

quasi.xs = read_xs(simple.quasi)
quasi.shift = data.frame(Station = "XS_800", horizontal = 5, 
  vertical = 2, stringsAsFactors = FALSE)
xs_shift(quasi.xs, "horizontal", quasi.shift)
xs_shift(quasi.xs, "horizontal", quasi.shift, subtract = FALSE)
xs_shift(quasi.xs, "vertical", quasi.shift)
xs_shift(quasi.xs, c("horizontal", "vertical"), quasi.shift)

}
