% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_volume}
\alias{xs_volume}
\title{Cross Section Volume}
\usage{
xs_volume(f, run.type, which.times = NULL, which.stations = NULL)
}
\arguments{
\item{f}{The h5 file to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of station numbers to extract. If
NULL, all stations will be returned.}
}
\description{
Compute cross section volumes from cross section data.
}
\details{
The boundaries of a cross section volume element are defined as
  halfway between the cross section and the next upstream and downstream
  cross sections. For cross sections at upstream or downstream boundaries of
  a reach a "half-volume" is computed, i.e. the cross section volume boundary
  is not extended beyond the reach boundaries.
}
