% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_volume_to_change}
\alias{xs_volume_to_change}
\title{Cross Section Volume Change}
\usage{
xs_volume_to_change(d, data.col = "Volume", station.col = "Station",
  time.col = "Time", region.col = "Region", region = NULL,
  cumulative = TRUE, longitudinal = TRUE)
}
\arguments{
\item{d}{The volume data.}

\item{data.col}{The column containing volume data.}

\item{station.col}{The column containing Station IDs.}

\item{time.col}{The column containing times.}

\item{region.col}{The column containing region IDs. Ignored if
\code{region} is \code{NULL}.}

\item{region}{The region to compute volume change for. If multiple
values are provided, the function will add the volumes of the specified
regions before computing change. Results will be erroneous if overlapping
regions are specified. If NULL, the function assumes the supplied
data is for a single region only. If \code{region} is  not named, the
function will assume the region column is named "Region". If \code{region}
is named, the name will be used to identify the column containing region
information.}

\item{cumulative}{If \code{TRUE}, accumulate volume change over time.}

\item{longitudinal}{If \code{TRUE}, accumulate volume change in the
 downstream direction. Only valid when all cross sections are included in
 \code{d}.
@return A wide data table containing volume change. If \code{over.time} or
  \code{longitudinal} are \code{TRUE}, data may be reordered.}
}
\description{
Compute cross-section volume change from volume data.
}
\details{
The \code{cumulative} and \code{longitudinal} arguments are simply
  flags to call \code{accumulate_data} on the output data using the same
  conventions as standard RAS output. If \code{cumulative} is \code{TRUE} and
  \code{longitudinal} is \code{FALSE}, the output data will essentially match
  Vol Bed Change Cum RAS output. If both \code{cumulative} and
  \code{longitudinal} are \code{TRUE}, the output data will essentially match
  the Long. Cum Vol Change RAS output.

 There may be some significant differences between the volume change computed
 by this function and the equivalent RAS output, depending on the cross
 section update tolerances specified in RAS and whether bank failure
 mechanisms (i.e. BSTEM) are incorporated into the model.
}
