% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{data_to_clipboard}
\alias{data_to_clipboard}
\title{Write Data To Clipboard}
\usage{
data_to_clipboard(d, header = TRUE)
}
\arguments{
\item{d}{The data table.}

\item{header}{If \code{TRUE}, write the column names to the first row.}
}
\value{
Writes the data table to the clipboard in tab-seperated format.
}
\description{
Write RAS data to clipboard in comma-seperated format for pasting
into e.g. Microsoft Excel.
}
