% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r, R/main.r
\name{diff_lsmd}
\alias{diff_lcmc}
\alias{diff_lsmd}
\alias{diff_mbcc}
\alias{diff_mic}
\alias{diff_moc}
\alias{diff_sediment}
\title{Difference Table (Sediment)}
\usage{
diff_lsmd(d1, d2)

diff_lcmc(d1, d2)

diff_mbcc(d1, d2)

diff_mic(d1, d2)

diff_moc(d1, d2)

diff_sediment(d1, d2, tcol, gcol, diffcol)
}
\arguments{
\item{d1}{The first dataframe.}

\item{d2}{The second dataframe.}

\item{tcol}{the time column name.}

\item{gcol}{the grain class column name.}

\item{diffcol}{The name of the difference column to be created.}
}
\value{
A dataframe.
}
\description{
Compute a difference table from sediment data.
}
\section{Functions}{
\itemize{
\item \code{diff_lsmd}: Compute a difference table for Lateral Structure Mass Divergence data.

\item \code{diff_lcmc}: Compute a difference table for Longitudinal Cumulative Mass Change data.

\item \code{diff_mbcc}: Compute a difference table for Mass Bed Change Cumulative data.

\item \code{diff_mic}: Compute a difference table for Mass In Cumulative data.

\item \code{diff_moc}: Compute a difference table for Mass Out Cumulative data.
}}

