% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{diff_sediment}
\alias{diff_sediment}
\title{Difference Table (Sediment)}
\usage{
diff_sediment(d1, d2, tcol, gcol, profcol, diffcol)
}
\arguments{
\item{d1}{The first dataframe.}

\item{d2}{The second dataframe.}

\item{tcol}{the time column name.}

\item{gcol}{the grain class column name.}

\item{profcol}{The name of the profile column to be created.}

\item{diffcol}{The name of the difference column to be created.}
}
\value{
A dataframe.
}
\description{
Compute a difference table from sediment data.
}

