% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{diff_sediment}
\alias{diff_sediment}
\alias{diff_fall_velocity}
\alias{diff_lat_struc_mass_div}
\alias{diff_long_cum_mass_change}
\alias{diff_long_cum_vol_change}
\alias{diff_long_cum_mass_moveable_limit}
\alias{diff_mass_bed_change}
\alias{diff_mass_bed_change_cum}
\alias{diff_mass_capacity}
\alias{diff_mass_cover}
\alias{diff_mass_in}
\alias{diff_mass_inactive}
\alias{diff_mass_in_cum}
\alias{diff_mass_out}
\alias{diff_mass_out_cum}
\alias{diff_mass_subsurface}
\alias{diff_reduce_armor_factor}
\title{Difference Table (Sediment)}
\usage{
diff_sediment(d1, d2, tcol, gcol, diffcol, percent = FALSE)

diff_fall_velocity(d1, d2)

diff_lat_struc_mass_div(d1, d2)

diff_long_cum_mass_change(d1, d2)

diff_long_cum_vol_change(d1, d2)

diff_long_cum_mass_moveable_limit(d1, d2)

diff_mass_bed_change(d1, d2)

diff_mass_bed_change_cum(d1, d2)

diff_mass_capacity(d1, d2)

diff_mass_cover(d1, d2)

diff_mass_in(d1, d2)

diff_mass_inactive(d1, d2)

diff_mass_in_cum(d1, d2)

diff_mass_out(d1, d2)

diff_mass_out_cum(d1, d2)

diff_mass_subsurface(d1, d2)

diff_reduce_armor_factor(d1, d2)
}
\arguments{
\item{d1}{The first dataframe, considered the "base" result.}

\item{d2}{The second dataframe, considered the "new" result.}

\item{tcol}{The time column name.}

\item{gcol}{the grain class column name.}

\item{diffcol}{The name of the difference column to be created.}

\item{percent}{Logical: report differences as percent difference.}
}
\value{
A dataframe in long table format, with difference defined as
 \code{d2- d1}. If \code{percent = TRUE}, the difference is defined
 as \code{(d2 - d1)/(0.5*(d2 + d1))}
}
\description{
Compute a difference table from sediment data.
}
\section{Functions}{
\itemize{
\item \code{diff_fall_velocity}: Compute a difference table for Fall Velocity data.

\item \code{diff_lat_struc_mass_div}: Compute a difference table for Lat Struc Mass Div data.

\item \code{diff_long_cum_mass_change}: Compute a difference table for Long. Cum Mass Change data.

\item \code{diff_long_cum_vol_change}: Compute a difference table for Long. Cum Vol Change data.

\item \code{diff_long_cum_mass_moveable_limit}: Compute a difference table for Long. Cum Mass Moveable Limit data.

\item \code{diff_mass_bed_change}: Compute a difference table for Mass Bed Change data.

\item \code{diff_mass_bed_change_cum}: Compute a difference table for Mass Bed Change Cum data.

\item \code{diff_mass_capacity}: Compute a difference table for Mass Capacity data.

\item \code{diff_mass_cover}: Compute a difference table for Mass Cover data.

\item \code{diff_mass_in}: Compute a difference table for Mass In data.

\item \code{diff_mass_inactive}: Compute a difference table for Mass Inactive data.

\item \code{diff_mass_in_cum}: Compute a difference table for Mass In Cum data.

\item \code{diff_mass_out}: Compute a difference table for Mass Out data.

\item \code{diff_mass_out_cum}: Compute a difference table for Mass Out Cum data.

\item \code{diff_mass_subsurface}: Compute a difference table for Mass Subsurface data.

\item \code{diff_reduce_armor_factor}: Compute a difference table for Reduce Armor Factor data.
}}

