% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{diff_table}
\alias{diff_d10_active}
\alias{diff_d10_cover}
\alias{diff_d10_inactive}
\alias{diff_d10_subsurface}
\alias{diff_d16_active}
\alias{diff_d16_cover}
\alias{diff_d16_inactive}
\alias{diff_d16_subsurface}
\alias{diff_d50_active}
\alias{diff_d50_cover}
\alias{diff_d50_inactive}
\alias{diff_d50_subsurface}
\alias{diff_d84_active}
\alias{diff_d84_cover}
\alias{diff_d84_inactive}
\alias{diff_d84_subsurface}
\alias{diff_d90_active}
\alias{diff_d90_cover}
\alias{diff_d90_inactive}
\alias{diff_d90_subsurface}
\alias{diff_dredged_cum}
\alias{diff_effective_depth}
\alias{diff_effective_width}
\alias{diff_flow}
\alias{diff_froude_number_channel}
\alias{diff_hydraulic_radius}
\alias{diff_invert_change}
\alias{diff_invert_elevation}
\alias{diff_mannings_n_channel}
\alias{diff_mean_effective_invert_change}
\alias{diff_mean_effective_invert_elevation}
\alias{diff_moveable_elv_l}
\alias{diff_moveable_elv_r}
\alias{diff_moveable_sta_l}
\alias{diff_moveable_sta_r}
\alias{diff_observed_data}
\alias{diff_sediment_concentration}
\alias{diff_shear_stress}
\alias{diff_shear_velocity}
\alias{diff_slope}
\alias{diff_table}
\alias{diff_temperature}
\alias{diff_thickness_cover}
\alias{diff_thickness_inactive}
\alias{diff_thickness_subsurface}
\alias{diff_velocity}
\alias{diff_water_surface}
\title{Difference Table}
\usage{
diff_table(d1, d2, tcol, diffcol, percent = FALSE)

diff_dredged_cum(d1, d2)

diff_effective_depth(d1, d2)

diff_effective_width(d1, d2)

diff_flow(d1, d2)

diff_froude_number_channel(d1, d2)

diff_hydraulic_radius(d1, d2)

diff_invert_change(d1, d2)

diff_invert_elevation(d1, d2)

diff_mannings_n_channel(d1, d2)

diff_mean_effective_invert_change(d1, d2)

diff_mean_effective_invert_elevation(d1, d2)

diff_moveable_elv_l(d1, d2)

diff_moveable_elv_r(d1, d2)

diff_moveable_sta_l(d1, d2)

diff_moveable_sta_r(d1, d2)

diff_observed_data(d1, d2)

diff_sediment_concentration(d1, d2)

diff_shear_stress(d1, d2)

diff_shear_velocity(d1, d2)

diff_slope(d1, d2)

diff_temperature(d1, d2)

diff_thickness_cover(d1, d2)

diff_thickness_inactive(d1, d2)

diff_thickness_subsurface(d1, d2)

diff_velocity(d1, d2)

diff_water_surface(d1, d2)

diff_d10_active(d1, d2)

diff_d10_cover(d1, d2)

diff_d10_inactive(d1, d2)

diff_d10_subsurface(d1, d2)

diff_d16_active(d1, d2)

diff_d16_cover(d1, d2)

diff_d16_inactive(d1, d2)

diff_d16_subsurface(d1, d2)

diff_d50_active(d1, d2)

diff_d50_cover(d1, d2)

diff_d50_inactive(d1, d2)

diff_d50_subsurface(d1, d2)

diff_d84_active(d1, d2)

diff_d84_cover(d1, d2)

diff_d84_inactive(d1, d2)

diff_d84_subsurface(d1, d2)

diff_d90_active(d1, d2)

diff_d90_cover(d1, d2)

diff_d90_inactive(d1, d2)

diff_d90_subsurface(d1, d2)
}
\arguments{
\item{d1}{The first dataframe, considered the "base" result.}

\item{d2}{The second dataframe, considered the "new" result.}

\item{tcol}{The time column name.}

\item{diffcol}{The name of the difference column to be created.}
}
\value{
A dataframe, with difference defined as \code{d2- d1}.
  if \code{percent = TRUE}, the difference is defined as
  \code{(d2 - d1)/(0.5*(d1 + d2))}
}
\description{
Compute a difference table.
}
\section{Functions}{
\itemize{
\item \code{diff_dredged_cum}: Compute a difference table for Dredged Cum data.

\item \code{diff_effective_depth}: Compute a difference table for Effective Depth data.

\item \code{diff_effective_width}: Compute a difference table for Effective_Width data.

\item \code{diff_flow}: Compute a difference table for Flow data.

\item \code{diff_froude_number_channel}: Compute a difference table for Froude Number Channel data.

\item \code{diff_hydraulic_radius}: Compute a difference table for Hydraulic Radius data.

\item \code{diff_invert_change}: Compute a difference table for Invert Change data.

\item \code{diff_invert_elevation}: Compute a difference table for Invert Elevation data.

\item \code{diff_mannings_n_channel}: Compute a difference table for Mannings n Channel data.

\item \code{diff_mean_effective_invert_change}: Compute a difference table for Mean Effective Invert Change data.

\item \code{diff_mean_effective_invert_elevation}: Compute a difference table for Mean Effective Invert Elevation data.

\item \code{diff_moveable_elv_l}: Compute a difference table for Moveable Elv L data.

\item \code{diff_moveable_elv_r}: Compute a difference table for Moveable Elv R data.

\item \code{diff_moveable_sta_l}: Compute a difference table for Moveable Sta L data.

\item \code{diff_moveable_sta_r}: Compute a difference table for Moveable Sta R data.

\item \code{diff_observed_data}: Compute a difference table for Observed Data data.

\item \code{diff_sediment_concentration}: Compute a difference table for Sediment Concentration data.

\item \code{diff_shear_stress}: Compute a difference table for Shear Stress data.

\item \code{diff_shear_velocity}: Compute a difference table for Shear Velocity data.

\item \code{diff_slope}: Compute a difference table for Slope data.

\item \code{diff_temperature}: Compute a difference table for Temperature data.

\item \code{diff_thickness_cover}: Compute a difference table for Thickness Cover data.

\item \code{diff_thickness_inactive}: Compute a difference table for Thickness Inactive data.

\item \code{diff_thickness_subsurface}: Compute a difference table for Thickness Subsurface data.

\item \code{diff_velocity}: Compute a difference table for Velocity data.

\item \code{diff_water_surface}: Compute a difference table for Water Surface data.

\item \code{diff_d10_active}: Compute a difference table for d10 Active data.

\item \code{diff_d10_cover}: Compute a difference table for d10 Cover data.

\item \code{diff_d10_inactive}: Compute a difference table for d10 Inactive data.

\item \code{diff_d10_subsurface}: Compute a difference table for d10 Subsurface data.

\item \code{diff_d16_active}: Compute a difference table for d16 Active data.

\item \code{diff_d16_cover}: Compute a difference table for d16 Cover data.

\item \code{diff_d16_inactive}: Compute a difference table for d16 Inactive data.

\item \code{diff_d16_subsurface}: Compute a difference table for d16 Subsurface data.

\item \code{diff_d50_active}: Compute a difference table for d50 Active data.

\item \code{diff_d50_cover}: Compute a difference table for d50 Cover data.

\item \code{diff_d50_inactive}: Compute a difference table for d50 Inactive data.

\item \code{diff_d50_subsurface}: Compute a difference table for d50 Subsurface data.

\item \code{diff_d84_active}: Compute a difference table for d84 Active data.

\item \code{diff_d84_cover}: Compute a difference table for d84 Cover data.

\item \code{diff_d84_inactive}: Compute a difference table for d84 Inactive data.

\item \code{diff_d84_subsurface}: Compute a difference table for d84 Subsurface data.

\item \code{diff_d90_active}: Compute a difference table for d90 Active data.

\item \code{diff_d90_cover}: Compute a difference table for d90 Cover data.

\item \code{diff_d90_inactive}: Compute a difference table for d90 Inactive data.

\item \code{diff_d90_subsurface}: Compute a difference table for d90 Subsurface data.
}}

