% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{difference_sediment}
\alias{difference_sediment}
\title{Difference Table (Sediment)}
\usage{
difference_sediment(d1, d2, relative = FALSE, partial = FALSE,
  difference.col = "Difference", time.col = "Time",
  grain.col = "GrainClass")
}
\arguments{
\item{d1}{The first dataframe, considered the "base" result.}

\item{d2}{The second dataframe, considered the "new" result.}

\item{relative}{Logical: report differences as relative difference.}

\item{partial}{If TRUE, only the overlapping times and columns will 
be processed.}

\item{difference.col}{The name of the difference column to be created.}

\item{time.col}{The time column name.}

\item{grain.col}{the grain class column name.}
}
\value{
A dataframe in long table format, with difference defined as
 \code{d2- d1}. If \code{relative = TRUE}, the difference is defined
 as \code{(d2 - d1)/(0.5*(d2 + d1))}
}
\description{
Compute a difference table from sediment data.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.volincum = read_sediment(simple.quasi, "Vol In Cum")

difference_sediment(quasi.volincum, quasi.volincum)

quasi.double = operate_sediment(quasi.volincum , fun = function(x) 2*x)
difference_sediment(quasi.volincum, quasi.double)
difference_sediment(quasi.volincum, quasi.double, relative = TRUE)

}
