% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{difference_table}
\alias{difference_table}
\title{Difference Table}
\usage{
difference_table(d1, d2, relative = FALSE, partial = FALSE,
  difference.col = "Difference", time.col = "Time")
}
\arguments{
\item{d1}{The first dataframe, considered the "base" result.}

\item{d2}{The second dataframe, considered the "new" result.}

\item{relative}{Logical: report differences as relative difference.}

\item{partial}{If TRUE, only the overlapping times and columns will 
be processed.}

\item{difference.col}{The name of the difference column to be created.}

\item{time.col}{The time column name.}
}
\value{
A dataframe, with difference defined as \code{d2- d1}.
  if \code{relative = TRUE}, the difference is defined as
  \code{(d2 - d1)/(0.5*(d2 + d1))}.
}
\description{
Compute a difference table.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.flow = read_standard(simple.quasi, "Flow")

difference_table(quasi.flow, quasi.flow)

quasi.double = operate_table(quasi.flow, fun = function(x) 2*x)
difference_table(quasi.flow, quasi.double)
difference_table(quasi.flow, quasi.double, relative = TRUE)

}
