% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{generate_report}
\alias{generate_report}
\title{Generate Report}
\usage{
generate_report(model1.file, model2.file, model1.type, model2.type,
  model1.label = NULL, model2.label = NULL, sections,
  sediment.table.classes, sediment.table.rows, output.file = "report",
  output.folder = tempdir(), output.html = TRUE)
}
\arguments{
\item{model1.file}{The path to the first RAS model HDF output.}

\item{model2.file}{The path to the second RAS model HDF output.}

\item{model1.type}{The type ("quasi" or "unsteady") of the first RAS model.}

\item{model2.type}{The type of the second RAS model.}

\item{sections}{The sections to include in the report.}

\item{sediment.table.classes}{If grain class sections are specified,
the grain classes to include in the report.}

\item{sediment.table.rows}{If grain class sections are specified,
which rows of the grain class tables to include in the report.}

\item{output.file}{The output filename, without an extension.}

\item{output.folder}{The folder to write out files to. Defaults to
the R temporary directory.}

\item{output.html}{If \code{TRUE}, generate an HTML report.}
}
\description{
Generate a RAStestR report.
}
\details{
Generating a report requires that the packages 
  \pkg{knitr}, \pkg{markdown} and \pkg{ggplot2}
  are installed. The following sections can be included in a report:
  \itemize{
    \item dredged cum 
    \item effective depth
    \item effective_width
    \item flow
    \item froude number channel
    \item invert change
    \item invert elevation
    \item lat struc mass div
    \item long. cum mass change
    \item mannings n channel
    \item mass bed change cum
    \item mass in cum
    \item mass out cum
    \item mean effective invert change
    \item mean effective invert elevation
    \item sediment concentration
    \item shear stress
    \item slope
    \item velocity
    \item water surface
    \item d10 active
    \item d10 inactive
    \item d50 active
    \item d50 inactive
    \item d90 active
    \item d90 inactive
  }
}
\seealso{
\link{read_standard}

\link{read_sediment}
}

