% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{generate_report}
\alias{generate_report}
\title{Generate Report}
\usage{
generate_report(model1.file, model2.file, model1.type, model2.type, sections,
  sediment.table.classes, sediment.table.rows, output.folder = "",
  output.html = TRUE)
}
\arguments{
\item{model1.file}{The path to the first RAS model HDF output.}

\item{model2.file}{The path to the second RAS model HDF output.}

\item{model1.type}{The type ("quasi" or "unsteady") of the first RAS model.}

\item{model2.type}{The type of the second RAS model.}

\item{sections}{The sections to include in the report.}

\item{sediment.table.classes}{If grain class sections are specified,
the grain classes to include in the report.}

\item{sediment.table.rows}{If grain class sections are specified,
which rows of the grain class tables to include in the report.}

\item{output.folder}{The folder to write out files to.}

\item{output.html}{If \code{TRUE}, generate an HTML report.}
}
\description{
Generate a RAStestR report.
}
\details{
Generating a report requires that the packages 
  \pkg{knitr}, \pkg{markdown} and \pkg{ggplot2}
  are installed. The following sections can be included in a report:
  \itemize{
    \item dredged cum
    \item effective depth
    \item effective_width
    \item flow
    \item invert change
    \item froude number channel
    \item invert change
    \item mannings n channel
    \item mean effective invert change
    \item mean effective invert elevation
    \item sediment concentration
    \item shear stress
    \item slope
    \item velocity
    \item water surface
    \item d10 active
    \item d10inactive
    \item d50active
    \item d50inactive
    \item d90 active
    \item d90 inactive
  }
}
\seealso{
\link{read_standard}

\link{read_sediment}
}

