% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.r
\name{generate_report}
\alias{generate_report}
\title{Generate Report}
\usage{
generate_report(model1.file, model2.file, model1.type, model2.type,
  model1.label = NULL, model2.label = NULL, sections,
  standard.opts = list(), sediment.opts = list(), output.file = "report",
  output.folder = tempdir(), output.type = c("html", "pdf"))
}
\arguments{
\item{model1.file}{The path to the first or "base" RAS model HDF output.}

\item{model2.file}{The path to the second or "new" RAS model HDF output.}

\item{model1.type}{The type ("quasi" or "unsteady") of the base RAS model.}

\item{model2.type}{The type of the new RAS model.}

\item{model1.label}{Optional label for the base RAS model.}

\item{model2.label}{Optional label for the new RAS model.}

\item{sections}{The sections to include in the report.}

\item{standard.opts}{List with elements \code{which.times} and 
\code{which.stations} specifying the times and stations to output
for standard tables. See \code{\link{read_standard}} for details.}

\item{sediment.opts}{List with elements \code{which.times},
\code{which.stations} and \code{which.grains} specifying the times,
stations and grain classes to output for sediment tables. See 
\code{\link{read_sediment}} for details.}

\item{output.file}{The output filename, without an extension.}

\item{output.folder}{The folder to write out files to. Defaults to
the R temporary directory.}

\item{output.type}{The file type to generate, either "html" or "pdf".}
}
\description{
Generate a RAStestR report.
}
\details{
Generating a report requires that the packages 
  \pkg{knitr}, \pkg{rmarkdown} and \pkg{ggplot2}
  are installed. The following sections can be included in a report:
  \itemize{
    \item dredged cum 
    \item effective depth
    \item effective_width
    \item flow
    \item froude number channel
    \item invert change
    \item invert elevation
    \item lat struc mass div
    \item long. cum mass change
    \item mannings n channel
    \item mass bed change cum
    \item mass in cum
    \item mass out cum
    \item mean effective invert change
    \item mean effective invert elevation
    \item sediment concentration
    \item shear stress
    \item slope
    \item velocity
    \item water surface
    \item d10 active
    \item d10 inactive
    \item d50 active
    \item d50 inactive
    \item d90 active
    \item d90 inactive
  }
}
\seealso{
\link{read_standard}

\link{read_sediment}
}

