% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_bank_stations}
\alias{list_bank_stations}
\title{List Bank Stations}
\usage{
list_bank_stations(f)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a vector of bank stations
}
\description{
Get list of bank stations
}
