% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_lengths}
\alias{list_lengths}
\title{List Cross Section Lengths}
\usage{
list_lengths(f)
}
\arguments{
\item{f}{The HDF5 file to read.}
}
\value{
a matrix of cross section lengths.
}
\description{
List cross section lengths.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_lengths(simple.quasi)

}
