% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_output_times}
\alias{list_output_times}
\title{List Time Steps}
\usage{
list_output_times(f, run.type)
}
\arguments{
\item{f}{The h5 file to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}
}
\value{
a vector of cross section lengths.
}
\description{
Get list of output times.
}
