% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{list_sediment}
\alias{list_sediment}
\title{List Sediment Tables}
\usage{
list_sediment(f, sedimentpath)
}
\arguments{
\item{f}{The h5 file to read.}
}
\value{
a vector of grain glass labels.
}
\description{
List grain class-specific tables of the specified type.
}
