% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_sediment}
\alias{list_sediment}
\title{List Sediment Tables}
\usage{
list_sediment(f, table.name)
}
\arguments{
\item{f}{The HDF5 file to read.}

\item{table.name}{The table to read.}
}
\value{
a vector of grain glass labels.
}
\description{
List grain class-specific tables of the specified type.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_sediment(simple.quasi, file.path("Results", "Sediment", 
    "Output Blocks", "Sediment", "Sediment Time Series", "Cross Sections", 
    "Vol In"))
RAStestR:::list_sediment(simple.quasi, file.path("Results", "Sediment", 
    "Output Blocks", "Sediment", "Sediment Time Series", "Cross Sections", 
    "Vol Inactive"))
RAStestR:::list_sediment(simple.quasi, file.path("Results", "Sediment", 
    "Output Blocks", "Sediment", "Sediment Time Series", "Cross Sections", 
    "Vol In Cum"))

}
