% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_stations}
\alias{list_stations}
\title{List River Stations}
\usage{
list_stations(f)
}
\arguments{
\item{f}{The hDF5 file to read.}
}
\value{
a vector of river station labels.
}
\description{
List RAS river station labels.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::list_stations(simple.quasi)

}
