% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablepaths.r
\name{list_xs}
\alias{list_xs}
\title{List Cross Section Tables}
\usage{
list_xs(f, xs.block)
}
\arguments{
\item{f}{The h5 file to read.}

\item{xs.block}{The HDF folder containing the cross section output tables.}
}
\value{
a vector of cross section output labels.
}
\description{
List cross section tables.
}
