% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{operate_table}
\alias{operate_table}
\title{Table Operations}
\usage{
operate_table(..., fun = "+", time.col = "Time")
}
\arguments{
\item{...}{Arbitrary number of wide-format data tables to combine.}

\item{fun}{A function to apply.}

\item{time.col}{The time column name.}
}
\value{
A single table.
}
\description{
Combine tables via an operation, e.g. addition or multiplication.
}
