% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{order_sediment}
\alias{order_sediment}
\title{Order Table (Sediment)}
\usage{
order_sediment(d, time.col = "Time", grain.col = "GrainClass")
}
\arguments{
\item{d}{A wide-format table.}

\item{time.col}{The time column name.}

\item{grain.col}{the grain class column name.}
}
\value{
the data frame \code{d}, ordered by time, cross section and grain class.
}
\description{
Reorder a sediment data table by time and cross section.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.voloutcum = read_sediment(simple.quasi, "Vol Out Cum")

quasi.disordered = quasi.voloutcum[sample(1:nrow(quasi.voloutcum), 
  nrow(quasi.voloutcum)),]
order_sediment(quasi.voloutcum)

}
