% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{read_bed_limits}
\alias{read_bed_limits}
\title{Read Bed Limits}
\usage{
read_bed_limits(f, run.type)
}
\arguments{
\item{f}{The h5 file to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}
}
\value{
a dataframe with columns "Time", "Station", "LOB" and "ROB", where
  columns "LOB" and "ROB" list the left and right bank stations,
  respectively.
}
\description{
Read the moveable bed limit stations.
}
\details{
The moveable bed limits will change over time in models that use
  BSTEM. In general, it is recommended but not required that the BSTEM toe
  stations are coincident with the moveable bed limits; otherwise, cross
  sections may adjust in unexpected ways.
}
