% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_hdtable}
\alias{read_hdtable}
\title{Read RAS Table}
\usage{
read_hdtable(f, table.path, row.table.path, col.table.path, rowcolname,
  colprefix)
}
\arguments{
\item{f}{The h5 file to read.}

\item{table.path}{The table to read in.}

\item{row.table.path}{The table containing the row identifiers.}

\item{col.table.path}{The table containing the column identifiers.}

\item{rowcolname}{The name to assign to the new row id column.}

\item{colprefix}{A prefix to apply to the column IDs.}
}
\value{
A dataframe.
}
\description{
Read RAS sediment data output. This function is used internally and should
not be called directly by the user.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
RAStestR:::read_hdtable(simple.quasi, 
  file.path("Results", "Sediment", "Output Blocks", "Sediment", 
    "Sediment Time Series", "Cross Sections", "Flow"),
  file.path("Results", "Sediment", "Output Blocks", "Sediment", 
    "Sediment Time Series", "Time Date Stamp"),
  file.path("Geometry", "Cross Sections", "River Stations"),
  "Time", "XS_")

}
