% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_hdtable}
\alias{read_hdtable}
\title{Generic Table Read Function}
\usage{
read_hdtable(f, tablepath, rowtablepath, coltablepath, run.type, rowcolname,
  colprefix)
}
\arguments{
\item{f}{The h5 file to read.}

\item{tablepath}{The table to read in.}

\item{rowtablepath}{The table containing the row identifiers.}

\item{coltablepath}{The table containing the column identifiers.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}

\item{rowcolname}{The name to assign to the new row id column.}

\item{colprefix}{A prefix to apply to the column IDs.}
}
\value{
A dataframe.
}
\description{
Read RAS sediment data output.
}

