% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{read_sediment}
\alias{read_sediment}
\alias{read_fall_velocity}
\alias{read_lat_struc_mass_div}
\alias{read_long_cum_mass_change}
\alias{read_long_cum_vol_change}
\alias{read_long_cum_mass_moveable_limit}
\alias{read_mass_bed_change}
\alias{read_mass_bed_change_cum}
\alias{read_mass_capacity}
\alias{read_mass_cover}
\alias{read_mass_in}
\alias{read_mass_inactive}
\alias{read_mass_in_cum}
\alias{read_mass_out}
\alias{read_mass_out_cum}
\alias{read_mass_subsurface}
\alias{read_reduce_armor_factor}
\title{Sediment By Grain Class Table}
\usage{
read_sediment(f, table.name, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_fall_velocity(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_lat_struc_mass_div(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_long_cum_mass_change(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_long_cum_vol_change(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_long_cum_mass_moveable_limit(f, which.times = NULL,
  which.stations = NULL, which.grains = NULL)

read_mass_bed_change(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_bed_change_cum(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_capacity(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_cover(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_in(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_inactive(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_in_cum(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_out(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_out_cum(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_mass_subsurface(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)

read_reduce_armor_factor(f, which.times = NULL, which.stations = NULL,
  which.grains = NULL)
}
\arguments{
\item{f}{The h5 file to read.}

\item{table.name}{The table to read.}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of stations to extract. If
NULL, all stations will be returned.}

\item{which.grains}{Grain class tables to extract. Can accept either numeric
grain class IDs or grain class labels. Label "ALL" or "" corresponds to
the totals. If NULL, all grain classes will be returned.}
}
\value{
A dataframe with a column "Time" containing the Date Time
  Stamp data; columns "XS_####" where ### is the cross-section ID;
  and column "GrainClass".
}
\description{
Read the sediment data output for all grain classes.
}
\section{Functions}{
\itemize{
\item \code{read_fall_velocity}: Read the Fall Velocity data output.

\item \code{read_lat_struc_mass_div}: Read the Lat Struc Mass Div data output.

\item \code{read_long_cum_mass_change}: Read the Long. Cum Mass Change data output.

\item \code{read_long_cum_vol_change}: Read the Long. Cum Vol Change data output.

\item \code{read_long_cum_mass_moveable_limit}: Read the Long. Cum Mass Moveable Limit data output.

\item \code{read_mass_bed_change}: Read the Mass Bed Change data output.

\item \code{read_mass_bed_change_cum}: Read the Mass Bed Change Cum data output.

\item \code{read_mass_capacity}: Read the Mass Capacity data output.

\item \code{read_mass_cover}: Read the Mass Cover data output.

\item \code{read_mass_in}: Read the Mass In data output.

\item \code{read_mass_inactive}: Read the Mass Inactive data output.

\item \code{read_mass_in_cum}: Read the Mass In Cum data output.

\item \code{read_mass_out}: Read the Mass Out data output.

\item \code{read_mass_out_cum}: Read the Mass Out Cum data output.

\item \code{read_mass_subsurface}: Read the Mass Subsurface data output.

\item \code{read_reduce_armor_factor}: Read the Reduce Armor Factor data output.
}}

\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
read_sediment(simple.quasi, "Vol Out Cum")

read_sediment(simple.quasi, "Vol Out Cum", 
  which.grains = c("6", "7"))
read_sediment(simple.quasi, "Vol Out Cum", 
  which.grains = c("VFS", "FS"))

}
