% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_sediment}
\alias{read_sediment}
\title{Sediment By Grain Class Table}
\usage{
read_sediment(f, table.name, which.times = NULL, which.stations = NULL,
  which.grains = NULL)
}
\arguments{
\item{f}{The hDF5 file to read.}

\item{table.name}{The table to read.}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of stations to extract. If
NULL, all stations will be returned.}

\item{which.grains}{Grain class tables to extract. Can accept either numeric
grain class IDs or grain class labels. Label "ALL" or "" corresponds to
the totals. If NULL, all grain classes will be returned.}
}
\value{
A dataframe with a column "Time" containing the Date Time
  Stamp data; columns "XS_####" where ### is the cross-section ID;
  and column "GrainClass".
}
\description{
Read the sediment data output for all grain classes.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
read_sediment(simple.quasi, "Vol Out Cum")

read_sediment(simple.quasi, "Vol Out Cum", 
  which.grains = c("6", "7"))
read_sediment(simple.quasi, "Vol Out Cum", 
  which.grains = c("VFS", "FS"))

}
