% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{read_standard}
\alias{read_d10a}
\alias{read_d10i}
\alias{read_d50a}
\alias{read_d50i}
\alias{read_d90a}
\alias{read_d90i}
\alias{read_dc}
\alias{read_ed}
\alias{read_ew}
\alias{read_f}
\alias{read_fnc}
\alias{read_ic}
\alias{read_ie}
\alias{read_meic}
\alias{read_meie}
\alias{read_mnc}
\alias{read_s}
\alias{read_sc}
\alias{read_ss}
\alias{read_standard}
\alias{read_v}
\alias{read_ws}
\title{Read Standard Table}
\usage{
read_standard(f, tablelab, run.type, which.times = NULL,
  which.stations = NULL)

read_dc(f, run.type, which.times = NULL, which.stations = NULL)

read_ed(f, run.type, which.times = NULL, which.stations = NULL)

read_ew(f, run.type, which.times = NULL, which.stations = NULL)

read_f(f, run.type, which.times = NULL, which.stations = NULL)

read_fnc(f, run.type, which.times = NULL, which.stations = NULL)

read_ic(f, run.type, which.times = NULL, which.stations = NULL)

read_ie(f, run.type, which.times = NULL, which.stations = NULL)

read_mnc(f, run.type, which.times = NULL, which.stations = NULL)

read_meic(f, run.type, which.times = NULL, which.stations = NULL)

read_meie(f, run.type, which.times = NULL, which.stations = NULL)

read_sc(f, run.type, which.times = NULL, which.stations = NULL)

read_ss(f, run.type, which.times = NULL, which.stations = NULL)

read_s(f, run.type, which.times = NULL, which.stations = NULL)

read_v(f, run.type, which.times = NULL, which.stations = NULL)

read_ws(f, run.type, which.times = NULL, which.stations = NULL)

read_d10a(f, run.type, which.times = NULL, which.stations = NULL)

read_d10i(f, run.type, which.times = NULL, which.stations = NULL)

read_d50a(f, run.type, which.times = NULL, which.stations = NULL)

read_d50i(f, run.type, which.times = NULL, which.stations = NULL)

read_d90a(f, run.type, which.times = NULL, which.stations = NULL)

read_d90i(f, run.type, which.times = NULL, which.stations = NULL)
}
\arguments{
\item{f}{The h5 file to read.}

\item{tablelab}{The table to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of station numbers to extract. If
NULL, all stations will be returned.}
}
\value{
A dataframe with a column "Time" containing the Date Time 
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read a standard (not grain class-specific) table.
}
\section{Functions}{
\itemize{
\item \code{read_dc}: Read the Dredged Cum data output.

\item \code{read_ed}: Read the Effective Depth data output.

\item \code{read_ew}: Read the Effective Width data output.

\item \code{read_f}: Read the Flow data output.

\item \code{read_fnc}: Read the Froude Number Channel data output.

\item \code{read_ic}: Read the Invert Change data output.

\item \code{read_ie}: Read the Invert Elevation data output.

\item \code{read_mnc}: Read the Manning's n Channel data output.

\item \code{read_meic}: Read the Mean Effective Invert Change data output.

\item \code{read_meie}: Read the Mean Effective Invert Elevation data output.

\item \code{read_sc}: Read the Sediment Concentration data output.

\item \code{read_ss}: Read the Shear Stress data output.

\item \code{read_s}: Read the Slope data output.

\item \code{read_v}: Read the Velocity data output.

\item \code{read_ws}: Read the Water Surface data output.

\item \code{read_d10a}: Read the d10 Active data output.

\item \code{read_d10i}: Read the d10 Inactive data output.

\item \code{read_d50a}: Read the d50 Active data output.

\item \code{read_d50i}: Read the d50 Inactive data output.

\item \code{read_d90a}: Read the d90 Active data output.

\item \code{read_d90i}: Read the d90 Inactive data output.
}}

