% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{read_standard}
\alias{read_standard}
\alias{read_dredged_cum}
\alias{read_effective_depth}
\alias{read_effective_width}
\alias{read_flow}
\alias{read_froude_number_channel}
\alias{read_hydraulic_radius}
\alias{read_invert_change}
\alias{read_invert_elevation}
\alias{read_mannings_n_channel}
\alias{read_mean_effective_invert_change}
\alias{read_mean_effective_invert_elevation}
\alias{read_moveable_elv_l}
\alias{read_moveable_elv_r}
\alias{read_moveable_sta_l}
\alias{read_moveable_sta_r}
\alias{read_observed_data}
\alias{read_sediment_concentration}
\alias{read_shear_stress}
\alias{read_shear_velocity}
\alias{read_slope}
\alias{read_temperature}
\alias{read_thickness_cover}
\alias{read_thickness_inactive}
\alias{read_thickness_subsurface}
\alias{read_velocity}
\alias{read_water_surface}
\alias{read_d10_active}
\alias{read_d10_cover}
\alias{read_d10_inactive}
\alias{read_d10_subsurface}
\alias{read_d16_active}
\alias{read_d16_cover}
\alias{read_d16_inactive}
\alias{read_d16_subsurface}
\alias{read_d50_active}
\alias{read_d50_cover}
\alias{read_d50_inactive}
\alias{read_d50_subsurface}
\alias{read_d84_active}
\alias{read_d84_cover}
\alias{read_d84_inactive}
\alias{read_d84_subsurface}
\alias{read_d90_active}
\alias{read_d90_cover}
\alias{read_d90_inactive}
\alias{read_d90_subsurface}
\title{Read Standard Table}
\usage{
read_standard(f, table.name, which.times = NULL, which.stations = NULL)

read_dredged_cum(f, which.times = NULL, which.stations = NULL)

read_effective_depth(f, which.times = NULL, which.stations = NULL)

read_effective_width(f, which.times = NULL, which.stations = NULL)

read_flow(f, which.times = NULL, which.stations = NULL)

read_froude_number_channel(f, which.times = NULL, which.stations = NULL)

read_hydraulic_radius(f, which.times = NULL, which.stations = NULL)

read_invert_change(f, which.times = NULL, which.stations = NULL)

read_invert_elevation(f, which.times = NULL, which.stations = NULL)

read_mannings_n_channel(f, which.times = NULL, which.stations = NULL)

read_mean_effective_invert_change(f, which.times = NULL,
  which.stations = NULL)

read_mean_effective_invert_elevation(f, which.times = NULL,
  which.stations = NULL)

read_moveable_elv_l(f, which.times = NULL, which.stations = NULL)

read_moveable_elv_r(f, which.times = NULL, which.stations = NULL)

read_moveable_sta_l(f, which.times = NULL, which.stations = NULL)

read_moveable_sta_r(f, which.times = NULL, which.stations = NULL)

read_observed_data(f, which.times = NULL, which.stations = NULL)

read_sediment_concentration(f, which.times = NULL, which.stations = NULL)

read_shear_stress(f, which.times = NULL, which.stations = NULL)

read_shear_velocity(f, which.times = NULL, which.stations = NULL)

read_slope(f, which.times = NULL, which.stations = NULL)

read_temperature(f, which.times = NULL, which.stations = NULL)

read_thickness_cover(f, which.times = NULL, which.stations = NULL)

read_thickness_inactive(f, which.times = NULL, which.stations = NULL)

read_thickness_subsurface(f, which.times = NULL, which.stations = NULL)

read_velocity(f, which.times = NULL, which.stations = NULL)

read_water_surface(f, which.times = NULL, which.stations = NULL)

read_d10_active(f, which.times = NULL, which.stations = NULL)

read_d10_cover(f, which.times = NULL, which.stations = NULL)

read_d10_inactive(f, which.times = NULL, which.stations = NULL)

read_d10_subsurface(f, which.times = NULL, which.stations = NULL)

read_d16_active(f, which.times = NULL, which.stations = NULL)

read_d16_cover(f, which.times = NULL, which.stations = NULL)

read_d16_inactive(f, which.times = NULL, which.stations = NULL)

read_d16_subsurface(f, which.times = NULL, which.stations = NULL)

read_d50_active(f, which.times = NULL, which.stations = NULL)

read_d50_cover(f, which.times = NULL, which.stations = NULL)

read_d50_inactive(f, which.times = NULL, which.stations = NULL)

read_d50_subsurface(f, which.times = NULL, which.stations = NULL)

read_d84_active(f, which.times = NULL, which.stations = NULL)

read_d84_cover(f, which.times = NULL, which.stations = NULL)

read_d84_inactive(f, which.times = NULL, which.stations = NULL)

read_d84_subsurface(f, which.times = NULL, which.stations = NULL)

read_d90_active(f, which.times = NULL, which.stations = NULL)

read_d90_cover(f, which.times = NULL, which.stations = NULL)

read_d90_inactive(f, which.times = NULL, which.stations = NULL)

read_d90_subsurface(f, which.times = NULL, which.stations = NULL)
}
\arguments{
\item{f}{The h5 file to read.}

\item{table.name}{The table to read.}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of stations to extract. If
NULL, all stations will be returned.}
}
\value{
A dataframe with a column "Time" containing the Date Time
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read a standard (not grain class-specific) table.
}
\section{Functions}{
\itemize{
\item \code{read_dredged_cum}: Read the Dredged Cum data output.

\item \code{read_effective_depth}: Read the Effective Depth data output.

\item \code{read_effective_width}: Read the Effective_Width data output.

\item \code{read_flow}: Read the Flow data output.

\item \code{read_froude_number_channel}: Read the Froude Number Channel data output.

\item \code{read_hydraulic_radius}: Read the Hydraulic Radius data output.

\item \code{read_invert_change}: Read the Invert Change data output.

\item \code{read_invert_elevation}: Read the Invert Elevation data output.

\item \code{read_mannings_n_channel}: Read the Mannings n Channel data output.

\item \code{read_mean_effective_invert_change}: Read the Mean Effective Invert Change data output.

\item \code{read_mean_effective_invert_elevation}: Read the Mean Effective Invert Elevation data output.

\item \code{read_moveable_elv_l}: Read the Moveable Elv L data output.

\item \code{read_moveable_elv_r}: Read the Moveable Elv R data output.

\item \code{read_moveable_sta_l}: Read the Moveable Sta L data output.

\item \code{read_moveable_sta_r}: Read the Moveable Sta R data output.

\item \code{read_observed_data}: Read the Observed Data data output.

\item \code{read_sediment_concentration}: Read the Sediment Concentration data output.

\item \code{read_shear_stress}: Read the Shear Stress data output.

\item \code{read_shear_velocity}: Read the Shear Velocity data output.

\item \code{read_slope}: Read the Slope data output.

\item \code{read_temperature}: Read the Temperature data output.

\item \code{read_thickness_cover}: Read the Thickness Cover data output.

\item \code{read_thickness_inactive}: Read the Thickness Inactive data output.

\item \code{read_thickness_subsurface}: Read the Thickness Subsurface data output.

\item \code{read_velocity}: Read the Velocity data output.

\item \code{read_water_surface}: Read the Water Surface data output.

\item \code{read_d10_active}: Read the d10 Active data output.

\item \code{read_d10_cover}: Read the d10 Cover data output.

\item \code{read_d10_inactive}: Read the d10 Inactive data output.

\item \code{read_d10_subsurface}: Read the d10 Subsurface data output.

\item \code{read_d16_active}: Read the d16 Active data output.

\item \code{read_d16_cover}: Read the d16 Cover data output.

\item \code{read_d16_inactive}: Read the d16 Inactive data output.

\item \code{read_d16_subsurface}: Read the d16 Subsurface data output.

\item \code{read_d50_active}: Read the d50 Active data output.

\item \code{read_d50_cover}: Read the d50 Cover data output.

\item \code{read_d50_inactive}: Read the d50 Inactive data output.

\item \code{read_d50_subsurface}: Read the d50 Subsurface data output.

\item \code{read_d84_active}: Read the d84 Active data output.

\item \code{read_d84_cover}: Read the d84 Cover data output.

\item \code{read_d84_inactive}: Read the d84 Inactive data output.

\item \code{read_d84_subsurface}: Read the d84 Subsurface data output.

\item \code{read_d90_active}: Read the d90 Active data output.

\item \code{read_d90_cover}: Read the d90 Cover data output.

\item \code{read_d90_inactive}: Read the d90 Inactive data output.

\item \code{read_d90_subsurface}: Read the d90 Subsurface data output.
}}

\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
read_standard(simple.quasi, "Flow")

simple.unsteady = system.file("sample-data/SampleUnsteady.hdf",
  package = "RAStestR")
read_standard(simple.unsteady, "Flow")

read_standard(simple.quasi, "Flow", which.times = "11DEC1990 01:00:00",
  which.stations = c("XS_800", "XS_796.00*"))

read_standard(simple.quasi, "Flow", which.times = 2:3,
  which.stations = 1:4)

}
