% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{read_standard}
\alias{read_ic}
\alias{read_standard}
\alias{read_ws}
\title{Read Standard Table}
\usage{
read_standard(f, tablelab, run.type)

read_ic(f, run.type)

read_ws(f, run.type)
}
\arguments{
\item{f}{The h5 file to read.}

\item{tablelab}{The table to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}
}
\value{
A dataframe with a column "Time" containing the Date Time 
  Stamp data and columns "XS_####" where ### is the cross-section ID.
}
\description{
Read a standard (not grain class-specific) table.
}
\section{Functions}{
\itemize{
\item \code{read_ic}: Read the Invert Change data output.

\item \code{read_ws}: Read the Water Surface data output.
}}

