% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{read_xs}
\alias{read_xs}
\title{Cross Section Table}
\usage{
read_xs(f, which.times = NULL, which.stations = NULL)
}
\arguments{
\item{f}{The h5 file to read.}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of stations to extract. If
NULL, all stations will be returned.}
}
\value{
A dataframe with a column "Time" containing the Date Time
  Stamp data; column "Station" containing the station ID in format "XS_####"
  where ### is the cross-section ID; column "Distance" containing the lateral
  distance location; and column "Elevation" containing the elevation at the
  distance location.
}
\description{
Read the cross section data output.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")

read_xs(simple.quasi) 
read_xs(simple.quasi, which.times = "02DEC1990 01:00:00",
  which.stations = c("XS_800", "XS_796.00*"))
read_xs(simple.quasi, which.times = 2:4, which.stations = 1:3)

}
