% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{reformat_fields}
\alias{reformat_fields}
\title{Reformat Standard Fields}
\usage{
reformat_fields(ld, fields = list("Time", "Station"))
}
\arguments{
\item{ld}{A data table.}

\item{fields}{The fields to reformat. For non-default Station and Time
columns, \code{fields} must be a named list identifying the actual names or
positions of the "Time" and "Station" columns.}
}
\value{
The data table with reformatted fields.
}
\description{
Reformat standard RAStestR fields as standard R data types.
}
\details{
If the "Time" field is specified, values in that column
  will be formatted as R timestamps (POSIXct). If the "Station"
  field is specified, stations will be converted to numeric values
  (only applicable to long format tables).
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.flow = read_standard(simple.quasi, "Flow")
reformat_fields(quasi.flow)

long.flow = to_longtable(quasi.flow, "Flow")
reformat_fields(long.flow)

}
