% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{rmse_table}
\alias{rmse_table}
\title{Root Mean Square Error Table}
\usage{
rmse_table(d, group.col, difference.col = "Difference", rmse.col = "RMSE")
}
\arguments{
\item{d}{The difference table.}

\item{group.col}{the column(s) to group differences by. For standard
tables, \code{group.col} will typically be either \code{"Station"}
 or \code{"Time"}. For sediment tables, \code{group.col} will
typically be either \code{c("GrainClass", "Station")} or
\code{c("GrainClass", "Time")}.}

\item{difference.col}{the column containing difference values.}

\item{rmse.col}{The output column containing RMSE values}
}
\value{
A dataframe.
}
\description{
Compute RMSE from a difference table.
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.flow = read_standard(simple.quasi, "Flow")
quasi.double = operate_table(quasi.flow, fun = function(x) 2*x)
quasi.difference = difference_table(quasi.flow, quasi.double)

rmse_table(quasi.difference, "Time", "Difference", "RMSE")
rmse_table(quasi.difference, "Station", "Difference", "RMSE")

quasi.volincum = read_sediment(simple.quasi, "Vol In Cum")
quasi.double = operate_sediment(quasi.volincum, fun = function(x) 2 * x)
quasi.difference = difference_sediment(quasi.volincum, quasi.double)
rmse_table(quasi.difference, c("Time", "GrainClass"), "Difference", "RMSE")
rmse_table(quasi.difference, c("Station", "GrainClass"), "Difference", "RMSE")

}
