% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{rmse_table}
\alias{rmse_table}
\title{Root Mean Square Error Table}
\usage{
rmse_table(d, groupcol, diffcol)
}
\arguments{
\item{d}{The difference table.}

\item{groupcol}{the column to group differences by.}

\item{diffcol}{the column containing difference values.}
}
\value{
A dataframe.
}
\description{
Compute RMSE from a difference table.
}

