% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r, R/main.r
\name{rmse_dc}
\alias{rmse_d10a}
\alias{rmse_d10i}
\alias{rmse_d50a}
\alias{rmse_d50i}
\alias{rmse_d90a}
\alias{rmse_d90i}
\alias{rmse_dc}
\alias{rmse_ed}
\alias{rmse_ew}
\alias{rmse_f}
\alias{rmse_fnc}
\alias{rmse_ic}
\alias{rmse_ie}
\alias{rmse_lcmc}
\alias{rmse_lsmd}
\alias{rmse_mbcc}
\alias{rmse_meic}
\alias{rmse_meie}
\alias{rmse_mic}
\alias{rmse_mnc}
\alias{rmse_moc}
\alias{rmse_s}
\alias{rmse_sc}
\alias{rmse_ss}
\alias{rmse_table}
\alias{rmse_v}
\alias{rmse_ws}
\title{Root Mean Square Error Table}
\usage{
rmse_dc(d, groupcol = "Station")

rmse_ed(d, groupcol = "Station")

rmse_ew(d, groupcol = "Station")

rmse_f(d, groupcol = "Station")

rmse_fnc(d, groupcol = "Station")

rmse_ic(d, groupcol = "Station")

rmse_ie(d, groupcol = "Station")

rmse_mnc(d, groupcol = "Station")

rmse_meic(d, groupcol = "Station")

rmse_meie(d, groupcol = "Station")

rmse_sc(d, groupcol = "Station")

rmse_ss(d, groupcol = "Station")

rmse_s(d, groupcol = "Station")

rmse_v(d, groupcol = "Station")

rmse_ws(d, groupcol = "Station")

rmse_d10a(d, groupcol = "Station")

rmse_d10i(d, groupcol = "Station")

rmse_d50a(d, groupcol = "Station")

rmse_d50i(d, groupcol = "Station")

rmse_d90a(d, groupcol = "Station")

rmse_d90i(d, groupcol = "Station")

rmse_lsmd(d, groupcol = "Station")

rmse_lcmc(d, groupcol = "Station")

rmse_mbcc(d, groupcol = "Station")

rmse_mic(d, groupcol = "Station")

rmse_moc(d, groupcol = "Station")

rmse_table(d, groupcol, diffcol, rmsecol)
}
\arguments{
\item{d}{The difference table.}

\item{groupcol}{the column(s) to group differences by. For standard
tables, \code{groupcol} will typically be either \code{"Station"}
 or \code{"Time"}. For sediment tables, \code{groupcol} will 
typically be either \code{c("GrainClass", "Station")} or
\code{c("GrainClass", "Time")}.}

\item{diffcol}{the column containing difference values.}

\item{rmsecol}{The output column containing RMSE values}
}
\value{
A dataframe.
}
\description{
Compute RMSE from a difference table.
}
\section{Functions}{
\itemize{
\item \code{rmse_dc}: Compute RMSE of Dredged Cumulative outputs.

\item \code{rmse_ed}: Compute RMSE of Effective Depth outputs.

\item \code{rmse_ew}: Compute RMSE of Effective Width outputs.

\item \code{rmse_f}: Compute RMSE of Flow outputs.

\item \code{rmse_fnc}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_ic}: Compute RMSE of Invert Change outputs.

\item \code{rmse_ie}: Compute RMSE of Invert Elevation outputs.

\item \code{rmse_mnc}: Compute RMSE of Manning's n Channel outputs.

\item \code{rmse_meic}: Compute RMSE of Mean Effective Invert Change outputs.

\item \code{rmse_meie}: Compute RMSE of Mean Effective Invert Elevation outputs.

\item \code{rmse_sc}: Compute RMSE of Sediment Concentration outputs.

\item \code{rmse_ss}: Compute RMSE of Shear Stress outputs.

\item \code{rmse_s}: Compute RMSE of Slope outputs.

\item \code{rmse_v}: Compute RMSE of Velocity outputs.

\item \code{rmse_ws}: Compute RMSE of Water Surface outputs.

\item \code{rmse_d10a}: Compute RMSE of d10 Active outputs.

\item \code{rmse_d10i}: Compute RMSE of d10 Inactive outputs.

\item \code{rmse_d50a}: Compute RMSE of d50 Active outputs.

\item \code{rmse_d50i}: Compute RMSE of d50 Inactive outputs.

\item \code{rmse_d90a}: Compute RMSE of d90 Active outputs.

\item \code{rmse_d90i}: Compute RMSE of d90 Inactive outputs.

\item \code{rmse_lsmd}: Compute RMSE of Lateral Structure Mass Divergence outputs.

\item \code{rmse_lcmc}: Compute RMSE of Longitudinal Cumulative Mass Change outputs.

\item \code{rmse_mbcc}: Compute RMSE of Mass Bed Change Cumulative outputs.

\item \code{rmse_mic}: Compute RMSE of Mass In Cumulative outputs.

\item \code{rmse_moc}: Compute RMSE of Mass Out Cumulative outputs.
}}

