% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r, R/helper.r
\name{rmse_table}
\alias{rmse_table}
\alias{rmse_dredged_cum}
\alias{rmse_effective_depth}
\alias{rmse_effective_width}
\alias{rmse_flow}
\alias{rmse_froude_number_channel}
\alias{rmse_hydraulic_radius}
\alias{rmse_invert_change}
\alias{rmse_invert_elevation}
\alias{rmse_mannings_n_channel}
\alias{rmse_mean_effective_invert_change}
\alias{rmse_mean_effective_invert_elevation}
\alias{rmse_moveable_elv_l}
\alias{rmse_moveable_elv_r}
\alias{rmse_moveable_sta_l}
\alias{rmse_moveable_sta_r}
\alias{rmse_observed_data}
\alias{rmse_sediment_concentration}
\alias{rmse_shear_stress}
\alias{rmse_shear_velocity}
\alias{rmse_slope}
\alias{rmse_temperature}
\alias{rmse_thickness_cover}
\alias{rmse_thickness_inactive}
\alias{rmse_thickness_subsurface}
\alias{rmse_velocity}
\alias{rmse_water_surface}
\alias{rmse_d10_active}
\alias{rmse_d10_cover}
\alias{rmse_d10_inactive}
\alias{rmse_d10_subsurface}
\alias{rmse_d16_active}
\alias{rmse_d16_cover}
\alias{rmse_d16_inactive}
\alias{rmse_d16_subsurface}
\alias{rmse_d50_active}
\alias{rmse_d50_cover}
\alias{rmse_d50_inactive}
\alias{rmse_d50_subsurface}
\alias{rmse_d84_active}
\alias{rmse_d84_cover}
\alias{rmse_d84_inactive}
\alias{rmse_d84_subsurface}
\alias{rmse_d90_active}
\alias{rmse_d90_cover}
\alias{rmse_d90_inactive}
\alias{rmse_d90_subsurface}
\alias{rmse_fall_velocity}
\alias{rmse_lat_struc_mass_div}
\alias{rmse_long_cum_mass_change}
\alias{rmse_long_cum_vol_change}
\alias{rmse_long_cum_mass_moveable_limit}
\alias{rmse_mass_bed_change}
\alias{rmse_mass_bed_change_cum}
\alias{rmse_mass_capacity}
\alias{rmse_mass_cover}
\alias{rmse_mass_in}
\alias{rmse_mass_inactive}
\alias{rmse_mass_in_cum}
\alias{rmse_mass_out}
\alias{rmse_mass_out_cum}
\alias{rmse_mass_subsurface}
\alias{rmse_reduce_armor_factor}
\title{Root Mean Square Error Table}
\usage{
rmse_table(d, group.col, difference.col = "Difference", rmse.col = "RMSE")

rmse_dredged_cum(d, group.col = "Station")

rmse_effective_depth(d, group.col = "Station")

rmse_effective_width(d, group.col = "Station")

rmse_flow(d, group.col = "Station")

rmse_froude_number_channel(d, group.col = "Station")

rmse_hydraulic_radius(d, group.col = "Station")

rmse_invert_change(d, group.col = "Station")

rmse_invert_elevation(d, group.col = "Station")

rmse_mannings_n_channel(d, group.col = "Station")

rmse_mean_effective_invert_change(d, group.col = "Station")

rmse_mean_effective_invert_elevation(d, group.col = "Station")

rmse_moveable_elv_l(d, group.col = "Station")

rmse_moveable_elv_r(d, group.col = "Station")

rmse_moveable_sta_l(d, group.col = "Station")

rmse_moveable_sta_r(d, group.col = "Station")

rmse_observed_data(d, group.col = "Station")

rmse_sediment_concentration(d, group.col = "Station")

rmse_shear_stress(d, group.col = "Station")

rmse_shear_velocity(d, group.col = "Station")

rmse_slope(d, group.col = "Station")

rmse_temperature(d, group.col = "Station")

rmse_thickness_cover(d, group.col = "Station")

rmse_thickness_inactive(d, group.col = "Station")

rmse_thickness_subsurface(d, group.col = "Station")

rmse_velocity(d, group.col = "Station")

rmse_water_surface(d, group.col = "Station")

rmse_d10_active(d, group.col = "Station")

rmse_d10_cover(d, group.col = "Station")

rmse_d10_inactive(d, group.col = "Station")

rmse_d10_subsurface(d, group.col = "Station")

rmse_d16_active(d, group.col = "Station")

rmse_d16_cover(d, group.col = "Station")

rmse_d16_inactive(d, group.col = "Station")

rmse_d16_subsurface(d, group.col = "Station")

rmse_d50_active(d, group.col = "Station")

rmse_d50_cover(d, group.col = "Station")

rmse_d50_inactive(d, group.col = "Station")

rmse_d50_subsurface(d, group.col = "Station")

rmse_d84_active(d, group.col = "Station")

rmse_d84_cover(d, group.col = "Station")

rmse_d84_inactive(d, group.col = "Station")

rmse_d84_subsurface(d, group.col = "Station")

rmse_d90_active(d, group.col = "Station")

rmse_d90_cover(d, group.col = "Station")

rmse_d90_inactive(d, group.col = "Station")

rmse_d90_subsurface(d, group.col = "Station")

rmse_fall_velocity(d, group.col = "Station")

rmse_lat_struc_mass_div(d, group.col = "Station")

rmse_long_cum_mass_change(d, group.col = "Station")

rmse_long_cum_vol_change(d, group.col = "Station")

rmse_long_cum_mass_moveable_limit(d, group.col = "Station")

rmse_mass_bed_change(d, group.col = "Station")

rmse_mass_bed_change_cum(d, group.col = "Station")

rmse_mass_capacity(d, group.col = "Station")

rmse_mass_cover(d, group.col = "Station")

rmse_mass_in(d, group.col = "Station")

rmse_mass_inactive(d, group.col = "Station")

rmse_mass_in_cum(d, group.col = "Station")

rmse_mass_out(d, group.col = "Station")

rmse_mass_out_cum(d, group.col = "Station")

rmse_mass_subsurface(d, group.col = "Station")

rmse_reduce_armor_factor(d, group.col = "Station")
}
\arguments{
\item{d}{The difference table.}

\item{group.col}{the column(s) to group differences by. For standard
tables, \code{group.col} will typically be either \code{"Station"}
 or \code{"Time"}. For sediment tables, \code{group.col} will
typically be either \code{c("GrainClass", "Station")} or
\code{c("GrainClass", "Time")}.}

\item{difference.col}{the column containing difference values.}

\item{rmse.col}{The output column containing RMSE values}
}
\value{
A dataframe.
}
\description{
Compute RMSE from a difference table.
}
\section{Functions}{
\itemize{
\item \code{rmse_dredged_cum}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_effective_depth}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_effective_width}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_flow}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_froude_number_channel}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_hydraulic_radius}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_invert_change}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_invert_elevation}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mannings_n_channel}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mean_effective_invert_change}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mean_effective_invert_elevation}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_moveable_elv_l}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_moveable_elv_r}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_moveable_sta_l}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_moveable_sta_r}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_observed_data}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_sediment_concentration}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_shear_stress}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_shear_velocity}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_slope}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_temperature}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_thickness_cover}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_thickness_inactive}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_thickness_subsurface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_velocity}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_water_surface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d10_active}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d10_cover}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d10_inactive}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d10_subsurface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d16_active}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d16_cover}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d16_inactive}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d16_subsurface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d50_active}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d50_cover}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d50_inactive}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d50_subsurface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d84_active}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d84_cover}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d84_inactive}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d84_subsurface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d90_active}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d90_cover}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d90_inactive}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_d90_subsurface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_fall_velocity}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_lat_struc_mass_div}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_long_cum_mass_change}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_long_cum_vol_change}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_long_cum_mass_moveable_limit}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_bed_change}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_bed_change_cum}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_capacity}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_cover}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_in}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_inactive}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_in_cum}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_out}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_out_cum}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_mass_subsurface}: Compute RMSE of Froude Number Channel outputs.

\item \code{rmse_reduce_armor_factor}: Compute RMSE of Froude Number Channel outputs.
}}

\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.flow = read_standard(simple.quasi, "Flow")
quasi.double = operate_table(quasi.flow, fun = function(x) 2*x)
quasi.difference = difference_table(quasi.flow, quasi.double)

rmse_table(quasi.difference, "Time", "Difference", "RMSE")
rmse_table(quasi.difference, "Station", "Difference", "RMSE")

quasi.volincum = read_sediment(simple.quasi, "Vol In Cum")
quasi.double = operate_sediment(quasi.volincum, fun = function(x) 2 * x)
quasi.difference = difference_sediment(quasi.volincum, quasi.double)
rmse_table(quasi.difference, c("Time", "GrainClass"), "Difference", "RMSE")
rmse_table(quasi.difference, c("Station", "GrainClass"), "Difference", "RMSE")

}
