% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{survey_interp}
\alias{survey_interp}
\title{Interpolate Survey Stations}
\usage{
survey_interp(d, by = c("space", "time"), time.col = "Time")
}
\arguments{
\item{d}{A wide-format table of processed data, e.g. output of
\code{xs_area}.}

\item{by}{Interpolate survey data over \code{space} or \code{time}.}

\item{time.col}{The time column name.}
}
\value{
The data frame \code{d} with some or all \code{NA} values filled in.
}
\description{
Interpolate data over stations. Useful for working with surveys that have
partial overlap.
}
\details{
If \code{by = "space"}, survey data will be interpolated linearly
  between adjacent cross sections at each separate time. If
  \code{by = "time"}, survey data will be interpolated linearly at each
  separate station. Note that in either case \code{d} will be reordered.
}
