% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{to_longtable}
\alias{to_longtable}
\title{Reformat As Long Table}
\usage{
to_longtable(d, data.col, station.col = "Station")
}
\arguments{
\item{d}{The data table, i.e. output from \code{read_table} or
\code{read_sediment}.}

\item{data.col}{The name of the new column holding data values.}

\item{station.col}{The name of the new column holding station IDs.
The prefix "XS_" will be stripped from the station IDs.}
}
\value{
The original data table in long format.
}
\description{
Reformat RAS data from wide table format (stations are individual
columns) to long table format (stations and data are collapsed to
key-value column format.
}
