% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.r
\name{to_widetable}
\alias{to_widetable}
\title{Reformat As Wide Table}
\usage{
to_widetable(d, key.col, value.col, key.prefix, key.attr = c("River",
  "Reach"))
}
\arguments{
\item{d}{The data table, i.e. output from \code{to_longtable}.}

\item{key.col}{The name of the column holding data keys.}

\item{value.col}{The name of the column holding data values.}

\item{key.prefix}{Text to prepend to the new columns created from
keys in \code{key.col}.}

\item{key.attr}{Columns to maintain as attributes in the resulting 
wide-format table. By default, columns "River" and "Reach" will be
searched for and maintained as attributes in the output wide-format}
}
\value{
The original data table in wide format.
}
\description{
Reformat RAS data from long table format (i.e. output of
\code{to_longtable}) to wide table format (key values are mapped to
individual columns).
}
\examples{
simple.quasi = system.file("sample-data/SampleQuasiUnsteady.hdf",
  package = "RAStestR")
quasi.volincum = read_sediment(simple.quasi, "Vol In Cum")
quasi.long = to_longtable(quasi.volincum, "VolIn")
to_widetable(quasi.long, "Station", "VolIn")
to_widetable(quasi.long, "GrainClass", "VolIn")

}
