% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_cumulative_change}
\alias{xs_cumulative_change}
\title{Cross Section Cumulative Change}
\usage{
xs_cumulative_change(d, time.col = "Time", station.col = "Station",
  distance.col = "Distance", elevation.col = "Elevation",
  bank.stations = NULL, reference.elevation = NULL,
  station.lengths = NULL, over.time = TRUE, longitudinal = TRUE,
  direction = "downstream")
}
\arguments{
\item{d}{The cross section data, i.e. output of \code{read_xs}.}

\item{time.col}{The column containing time stamps.}

\item{station.col}{The column containing cross section station IDs.}

\item{distance.col}{The column containing cross section distances.}

\item{elevation.col}{The column containing cross section elevations.}

\item{bank.stations}{A table defining the bank station distances to use when
computing cross section area. If \code{NULL}, the full extent of each cross
section will be used. See details for more information.}

\item{reference.elevation}{Use predefined reference elevations when computing
cross section area. Can be a constant or a two column station of cross
  section labels and associated reference elevation. If \code{NULL}, the
  maximum elevation of each cross section will be used.}

\item{station.lengths}{The cross section length values.}

\item{over.time}{If \code{TRUE}, accumulate data across time steps. This
is generally valid only for data output at the computation time step.}

\item{longitudinal}{If \code{TRUE}, accumulate data along the reach. This
is generally only valid when all cross sections are included in \code{d}.}

\item{direction}{Accumulate data in the downstream (descending order of cross
section IDs) or upstream (ascending order) direction. Ignored if
\code{longitudinal} is \code{FALSE}.}
}
\description{
Compute cumulative cross section change directy from cross section data.
}
\details{
This is essentially a wrapper for processing cross section data
  through the sequence \code{xs_area} --> \code{area_to_volume} -->
  \code{change_table} --> \code{cumulative_table.}
}
