% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_lengths}
\alias{xs_lengths}
\title{Compute Cross Section Lengths}
\usage{
xs_lengths(d, station.list = NULL)
}
\arguments{
\item{d}{A data frame with columns "Station", "LOB", "Channel" and "ROB",
i.e. output of \code{read_station_lengths}.}

\item{station.list}{Optional list of stations to extract. If stations are}
}
\value{
a dataframe with columns "Station" and "Length", where column
"Length" lists the cross section length.
}
\description{
Compute cross section lengths from the river station lengths table.
}
\details{
The cross section length is computed as the average of of the
  current river station channel length and the upstream river station channel
  length; in other words, the cross section control volume is assumed to
  extend half the channel distance between the next upstream and downstream
  cross sections.
}
