% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_volume}
\alias{xs_volume}
\title{Cross Section Volume}
\usage{
xs_volume(f, run.type, which.times = NULL, which.stations = NULL,
  which.regions = c("total", "bed", "banks", "left bank", "right bank"),
  fix.bank.stations = FALSE)
}
\arguments{
\item{f}{The h5 file to read.}

\item{run.type}{The model run type, e.g. "quasi" or "unsteady".}

\item{which.times}{Character vector of timestamps to extract. If
NULL, all timestamps will be returned.}

\item{which.stations}{Character vector of station numbers to extract. If
NULL, all stations will be returned.}

\item{which.regions}{List of sections of the cross section to compute volume for.
See details section for more information.}

\item{fix.bank.stations}{If \code{TRUE}, the separation between "bed" and
"bank" regions will be defined by the starting moveable bed limits. If
\code{FALSE}, the separation between regions vary through time according to
 values in the Moveable Bed Station tables.}
}
\value{
A data frame with columns "Time", "Station", "Volume" and "Region".
}
\description{
Compute cross section volumes from cross section data.
}
\details{
The boundaries of a cross section volume element are defined as
  halfway between the cross section and the next upstream and downstream
  cross sections. For cross sections at upstream or downstream boundaries of
  a reach a "half-volume" is computed, i.e. the cross section volume boundary
  is not extended beyond the reach boundaries.

  Volume change can be computed for separate sections of the cross section
  through the argument \code{which.regions}. The following choices are available:
  \describe{
    \item{"total"}{Compute the volume of the entire cross section.}
    \item{"bed"}{Compute bed change volume only. The bed is defined as the region spanned by the moveable bed limits.}
    \item{"banks"}{Compute the bank change volume only. The banks are defined by the region outside of the moveable bed limits.}
    \item{"left bank"}{Compute the left bank change volume only.}
    \item{"right bank"}{Compute the right bank change volume only.}
  }
  Note that the section "banks" is equal to the sum of the "left bank" and
  "right bank" sections, and "total" is equal to the sum of "bed" and
  "banks".
}
