% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_volume_to_change}
\alias{xs_volume_to_change}
\title{Cross Section Volume Change}
\usage{
xs_volume_to_change(d, station.col = "Station", time.col = "Time",
  cumulative = TRUE, longitudinal = TRUE)
}
\arguments{
\item{d}{The volume data.}

\item{station.col}{The column containing Station IDs.}

\item{time.col}{The column containing times.}

\item{cumulative}{If \code{TRUE}, accumulate volume change over time.}

\item{longitudinal}{If \code{TRUE}, accumulate volume change in the
 downstream direction. Only valid when all cross sections are included in
 \code{d}.
@return A wide data table containing volume change. If \code{over.time} or
  \code{longitudinal} are \code{TRUE}, data may be reordered.}
}
\description{
Compute cross-section volume change from volume data.
}
