% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xs.r
\name{xs_volume_to_change}
\alias{xs_volume_to_change}
\title{Cross Section Volume Change}
\usage{
xs_volume_to_change(d, station.col = "Station", time.col = "Time",
  cumulative = TRUE, longitudinal = TRUE)
}
\arguments{
\item{d}{The volume data.}

\item{station.col}{The column containing Station IDs.}

\item{time.col}{The column containing times.}

\item{cumulative}{If \code{TRUE}, accumulate volume change over time.}

\item{longitudinal}{If \code{TRUE}, accumulate volume change in the
 downstream direction. Only valid when all cross sections are included in
 \code{d}.
@return A wide data table containing volume change. If \code{over.time} or
  \code{longitudinal} are \code{TRUE}, data may be reordered.}
}
\description{
Compute cross-section volume change from volume data.
}
\details{
The \code{cumulative} and \code{longitudinal} arguments are simply
  flags to call \code{accumulate_data} on the output data using the same
  conventions as standard RAS output. If \code{cumulative} is \code{TRUE} and
  \code{longitudinal} is \code{FALSE}, the output data will essentially match
  Vol Bed Change Cum RAS output. If both \code{cumulative} and
  \code{longitudinal} are \code{TRUE}, the output data will essentially match
  the Long. Cum Vol Change RAS output.

 There may be some significant differences between the volume change computed
 by this function and the equivalent RAS output, depending on the cross
 section update tolerances specified in RAS and whether bank failure
 mechanisms (i.e. BSTEM) are incorporated into the model.
}
